% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alink.R
\name{alink}
\alias{alink}
\title{Return a Link To Download an Artifact Stored on GitHub Repository}
\usage{
alink(md5hash, repo = aoptions("repo"), user = aoptions("repo"),
  format = "markdown", rawLink = FALSE)
}
\arguments{
\item{md5hash}{A character assigned to the artifact through the use of a cryptographical hash function with MD5 algorithm. See more \link{md5hash}. 
If specified in a format of \code{'repo/user/md5hash'} then the \code{user} and \code{repo} parameters are omitted.}

\item{repo}{The GitHub \code{Repository} on which the desired to be download artifact
is stored.}

\item{user}{The name of a user on whos \code{Repository} the desired to be download artifact 
is stored.}

\item{format}{In which format the link should be returned. Possibilites are \code{markdown} (default) or \code{latex}.}

\item{rawLink}{A logical denoting whether to return raw link or a link in \code{format} convention.}
}
\value{
This function returns a link to download artifact that are archived on GitHub.
}
\description{
\code{alink} returns a link to download and artifact from GitHub \link{Repository}.
Artifact has to be prior archived on github, e.g with \link{archive} function (recomended) or 
\link{saveToRepo} function and traditional git manual synchronization.
To learn more about artifacts visit \link[archivist]{archivist-package}.
}
\details{
For more information about \code{md5hash} see \link{md5hash}.
}
\examples{
\dontrun{


alink('pbiecek/archivist/134ecbbe2a8814d98f0c2758000c408e')
alink(user = 'MarcinKosinski', repo = 'Museum',
 md5hash = '1651caa499a2b07a3bdad3896a2fc717', format = 'latex)
alink('pbiecek/graphGallery/02af4f99e440324b9e329faa293a9394', rawLink = TRUE)  

## Exmaple with archive function and 
## empty Github Repository creation

library(httr)
myapp <- oauth_app("github",
                   key = app_key,
                   secret = app_secret)
github_token <- oauth2.0_token(oauth_endpoints("github"),
                               myapp,
                               scope = "public_repo")
# setting options                              
aoptions("github_token", github_token)
aoptions("user.name", user_name)
aoptions("user.password", user_password)

createEmptyGithubRepo("archive-test4", default = TRUE)
## artifact's archiving
vectorLong <- 1:100
vectorShort <- 1:20
# archiving
alink(archive(vectorLong))
archive(vectorShort, alink = TRUE)
}
}
\author{
Marcin Kosinski, \email{m.p.kosinski@gmail.com}
}
\seealso{
Other archivist: \code{\link{Repository}},
  \code{\link{Tags}}, \code{\link{\%a\%}},
  \code{\link{addTagsRepo}}, \code{\link{ahistory}},
  \code{\link{aoptions}}, \code{\link{archive}},
  \code{\link{archivist-github-integration}},
  \code{\link{archivist-package}}, \code{\link{aread}},
  \code{\link{asearch}}, \code{\link{cache}},
  \code{\link{cloneGithubRepo}},
  \code{\link{copyLocalRepo}},
  \code{\link{createEmptyRepo}}, \code{\link{deleteRepo}},
  \code{\link{getTagsLocal}},
  \code{\link{loadFromLocalRepo}}, \code{\link{md5hash}},
  \code{\link{rmFromRepo}}, \code{\link{saveToRepo}},
  \code{\link{searchInLocalRepo}},
  \code{\link{setLocalRepo}},
  \code{\link{shinySearchInLocalRepo}},
  \code{\link{showLocalRepo}},
  \code{\link{splitTagsLocal}},
  \code{\link{summaryLocalRepo}},
  \code{\link{zipLocalRepo}}
}

