% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{archivist-package}
\alias{archivist-package}
\title{Tools for archivisation of R objects.}
\description{
A description.
}
\details{
To be written.
}
\author{
authors1 [aut] \email{mails} \cr
authors2 [aut, cre] \email{mails}
}
\seealso{
Other archivist: \code{\link{Tags}};
  \code{\link{createEmptyRepo}};
  \code{\link{loadFromGithubRepo}},
  \code{\link{loadFromLocalRepo}};
  \code{\link{rmFromRepo}}; \code{\link{saveToRepo}};
  \code{\link{searchInGithubRepo}},
  \code{\link{searchInLocalRepo}}
}

