% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/archivistOptions.R
\name{setArchivistOptions}
\alias{setArchivistOptions}
\title{Default options for Archivist}
\usage{
setArchivistOptions(key, value)
}
\arguments{
\item{key}{Name of the parameter.}

\item{value}{New value of the 'key' parameter.}
}
\value{
Both functions return value that corresponds to a selected key.
}
\description{
\code{setArchivistOptions} and \code{getArchivistOptions} set and get default opitions
for many \code{archivist} functions.
}
\details{
\code{setArchivistOptions} function sets a parameter that will be stored in internal environment.
\code{getArchivistOptions} returns value of given parameter.
Both functions are used for setting default values for parameters in archivist functions.
Currently works for arguments: silent
}
\examples{
# objects preparation
\dontrun{
# turn off warnings in saveToRepo()
setArchivistOptions("silent", FALSE)
}
}
\author{
Przemysław Biecek, \email{przemyslaw.biecek@gmail.com}
}
\seealso{
Other archivist: \code{\link{Repository}};
  \code{\link{Tags}}; \code{\link{\%a\%}};
  \code{\link{addTagsRepo}}; \code{\link{ahistory}},
  \code{\link{print.ahistory}};
  \code{\link{archivist-package}}; \code{\link{aread}};
  \code{\link{asearch}}; \code{\link{cache}};
  \code{\link{copyGithubRepo}},
  \code{\link{copyLocalRepo}};
  \code{\link{createEmptyRepo}}; \code{\link{deleteRepo}};
  \code{\link{getTagsGithub}}, \code{\link{getTagsLocal}};
  \code{\link{loadFromGithubRepo}},
  \code{\link{loadFromLocalRepo}}; \code{\link{md5hash}};
  \code{\link{multiSearchInGithubRepo}},
  \code{\link{multiSearchInLocalRepo}},
  \code{\link{searchInGithubRepo}},
  \code{\link{searchInLocalRepo}};
  \code{\link{rmFromRepo}}; \code{\link{saveSetToRepo}};
  \code{\link{saveToRepo}}; \code{\link{setGithubRepo}},
  \code{\link{setLocalRepo}};
  \code{\link{shinySearchInLocalRepo}};
  \code{\link{showGithubRepo}},
  \code{\link{showLocalRepo}};
  \code{\link{summaryGithubRepo}},
  \code{\link{summaryLocalRepo}};
  \code{\link{zipGithubRepo}}, \code{\link{zipLocalRepo}}
}

