% Generated by roxygen2 (4.0.1): do not edit by hand
\name{copyLocalRepo}
\alias{copyGithubRepo}
\alias{copyLocalRepo}
\title{Copy an Existing Repository to Another Repository}
\usage{
copyLocalRepo(repoFrom, repoTo, md5hashes)

copyGithubRepo(repoTo, md5hashes, user, repo, branch = "master")
}
\arguments{
\item{repoFrom}{A character that specifies the directory of the Repository from which
artifacts will be copied. Works only on \code{copyLocalRepo}.}

\item{repoTo}{A character that specifies the directory of the Repository in which
artifacts will be copied.}

\item{md5hashes}{A character or character vector containing \code{md5hashes} of artifacts to be copied.}

\item{repo}{Only if coping a Github repository. A character containing a name of a Github repository on which the \code{repoFrom}-Repository is archived.}

\item{user}{Only if coping a Github repository. A character containing a name of a Github user on whose account the \code{repoFrom} is created.}

\item{branch}{Only if coping with a Github repository. A character containing a name of
Github Repository's branch on which a \code{repoFrom}-Repository is archived. Default \code{branch} is \code{master}.}
}
\description{
\code{copyToRepo} copies artifact from one \link{Repository} to another \code{Repository}.
Functions \code{copyLocalRepo} and \code{copyGithubRepo} copy artifacts from the archivist Repositories stored in a local folder or on Github.
Both of them take \code{md5hash} as a parameter, which is a result from \link{saveToRepo} function.
For every artifacts, \code{md5hash} is a unique string of length 32 that comes out as a result of
\link[digest]{digest} function, which uses a cryptographical MD5 hash algorithm. For more information see \link{md5hash}.
}
\details{
\code{copyToRepo} copies artifact from one \code{Repository} to another \code{Repository}. It addes new files
to exising \code{gallery} folder in \code{repoTo} \code{Repository}. \code{copyLocalRepo} copies local \code{Repository}, where
\code{copyGithubRepo} copies Github \code{Repository}.
}
\examples{
\dontrun{


# creating example Repository - that examples will work

exampleRepoDir <- tempdir()
hashes <- searchInGithubRepo( pattern="name", user="pbiecek", repo="archivist", fixed=FALSE )

createEmptyRepo( exampleRepoDir )

copyGithubRepo( repoTo = exampleRepoDir , md5hashes= hashes, user="pbiecek", repo="archivist" )

# removing an example Repository

deleteRepo( exampleRepoDir )

rm( exampleRepoDir )


}
}
\author{
Marcin Kosinski, \email{m.p.kosinski@gmail.com}
}
\seealso{
Other archivist: \code{\link{Repository}};
  \code{\link{Tags}}; \code{\link{archivist-package}};
  \code{\link{createEmptyRepo}}; \code{\link{deleteRepo}};
  \code{\link{loadFromGithubRepo}},
  \code{\link{loadFromLocalRepo}}; \code{\link{md5hash}};
  \code{\link{returnTag}}; \code{\link{rmFromRepo}};
  \code{\link{saveToRepo}};
  \code{\link{searchInGithubRepo}},
  \code{\link{searchInLocalRepo}};
  \code{\link{showGithubRepo}},
  \code{\link{showLocalRepo}};
  \code{\link{summaryGithubRepo}},
  \code{\link{summaryLocalRepo}}
}

