% Generated by roxygen2 (4.0.1): do not edit by hand
\name{createEmptyRepo}
\alias{createEmptyRepo}
\title{Create an Empty Repository in Given Directory}
\usage{
createEmptyRepo(dir)
}
\arguments{
\item{dir}{A character string that specifies the directory for repository to be made.}
}
\description{
\code{createEmptyRepo} creates an empty repository for given directory in which archivised objects will be managed.
}
\details{
This function must be initialized prior to use archivist package.
If working in groups, it is higly recommend to create repository on shared dropbox folder.
}
\examples{
createEmptyRepo(getwd())
createEmptyRept(path.package("stats"))
# not work
createEmptyRepo("user/folder/here")
}
\author{
autor
}
\seealso{
Other archivist: \code{\link{loadFromGithubRepo}},
  \code{\link{loadFromLocalRepo}};
  \code{\link{rmFromRepo}}; \code{\link{saveToRepo}};
  \code{\link{searchInGithubRepo}},
  \code{\link{searchInLocalRepo}}
}

