% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractData.R
\name{extractData}
\alias{extractData}
\alias{extractData.default}
\alias{extractData.ggplot}
\alias{extractData.lm}
\alias{extractData.htest}
\alias{extractData.lda}
\alias{extractData.trellis}
\alias{extractData.twins}
\alias{extractData.partition}
\alias{extractData.qda}
\alias{extractData.glmnet}
\alias{extractData.survfit}
\title{Internal function for extraction of data from objects}
\usage{
extractData(object, parrentMd5hash, parentDir, isForce, ASCII)

\method{extractData}{default}(object, parrentMd5hash, parentDir, isForce, ASCII)

\method{extractData}{ggplot}(object, parrentMd5hash, parentDir, isForce, ASCII)

\method{extractData}{lm}(object, parrentMd5hash, parentDir, isForce, ASCII)

\method{extractData}{htest}(object, parrentMd5hash, parentDir, isForce, ASCII)

\method{extractData}{lda}(object, parrentMd5hash, parentDir, isForce, ASCII)

\method{extractData}{trellis}(object, parrentMd5hash, parentDir, isForce, ASCII)

\method{extractData}{twins}(object, parrentMd5hash, parentDir, isForce, ASCII)

\method{extractData}{partition}(object, parrentMd5hash, parentDir, isForce, ASCII)

\method{extractData}{qda}(object, parrentMd5hash, parentDir, isForce, ASCII)

\method{extractData}{glmnet}(object, parrentMd5hash, parentDir, isForce, ASCII)

\method{extractData}{survfit}(object, parrentMd5hash, parentDir, isForce, ASCII)
}
\arguments{
\item{object}{for this object data is to be extracted}

\item{parrentMd5hash}{hash of the parent object}

\item{parentDir}{parent dir}

\item{isForce}{should the data extraction be forced}

\item{ASCII}{shall it be written in ASCII friendly format}
}
\value{
\code{NULL}

\code{NULL}

\code{NULL}

\code{NULL}

\code{NULL}

\code{NULL}

\code{NULL}

\code{NULL}

\code{NULL}

\code{NULL}

\code{NULL}
}
\description{
Internal function for extraction of data from objects
}
