% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMiniature.R
\name{extractMiniature}
\alias{extractMiniature}
\alias{extractMiniature.default}
\alias{extractMiniature.data.frame}
\alias{extractMiniature.ggplot}
\alias{extractMiniature.recordedplot}
\alias{extractMiniature.lm}
\alias{extractMiniature.htest}
\alias{extractMiniature.trellis}
\alias{extractMiniature.twins}
\alias{extractMiniature.partition}
\alias{extractMiniature.lda}
\alias{extractMiniature.qda}
\alias{extractMiniature.glmnet}
\alias{extractMiniature.survfit}
\title{Internal function for extraction of miniatures}
\usage{
extractMiniature(object, md5hash, parentDir, ...)

\method{extractMiniature}{default}(object, md5hash, parentDir, ...)

\method{extractMiniature}{data.frame}(object, md5hash, parentDir, ..., firstRows = 6)

\method{extractMiniature}{ggplot}(object, md5hash, parentDir, ..., width = 800, height = 600)

\method{extractMiniature}{recordedplot}(object, md5hash, parentDir, ..., width = 800, height = 600)

\method{extractMiniature}{lm}(object, md5hash, parentDir, ...)

\method{extractMiniature}{htest}(object, md5hash, parentDir, ...)

\method{extractMiniature}{trellis}(object, md5hash, parentDir, ..., width = 800, height = 600)

\method{extractMiniature}{twins}(object, md5hash, parentDir, ..., width = 800, height = 600)

\method{extractMiniature}{partition}(object, md5hash, parentDir, ..., width = 800, height = 600)

\method{extractMiniature}{lda}(object, md5hash, parentDir, ...)

\method{extractMiniature}{qda}(object, md5hash, parentDir, ...)

\method{extractMiniature}{glmnet}(object, md5hash, parentDir, ...)

\method{extractMiniature}{survfit}(object, md5hash, parentDir, ...)
}
\arguments{
\item{object}{for this object miniatures are extracted}

\item{md5hash}{hash of the object}

\item{parentDir}{parent dir}

\item{...}{other arguments}

\item{firstRows}{what to present in the miniature}

\item{width}{how large should be the miniature?}

\item{height}{how large should be the miniature?}
}
\value{
\code{NULL}

\code{NULL}

\code{NULL}

\code{NULL}

\code{NULL}

\code{NULL}

\code{NULL}

\code{NULL}

\code{NULL}

\code{NULL}

\code{NULL}

\code{NULL}

\code{NULL}
}
\description{
Internal function for extraction of miniatures
}
