% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractTags.R
\name{extractTags}
\alias{extractTags}
\alias{extractTags.default}
\alias{extractTags.data.frame}
\alias{extractTags.ggplot}
\alias{extractTags.trellis}
\alias{extractTags.lm}
\alias{extractTags.summary.lm}
\alias{extractTags.htest}
\alias{extractTags.lda}
\alias{extractTags.qda}
\alias{extractTags.twins}
\alias{extractTags.partition}
\alias{extractTags.glmnet}
\alias{extractTags.survfit}
\title{Internal function for extraction of tags from objects}
\usage{
extractTags(object, objectNameX, ...)

\method{extractTags}{default}(object, objectNameX, ...)

\method{extractTags}{data.frame}(object, objectNameX, ...)

\method{extractTags}{ggplot}(object, objectNameX, ...)

\method{extractTags}{trellis}(object, objectNameX, ...)

\method{extractTags}{lm}(object, objectNameX, ...)

\method{extractTags}{summary.lm}(object, objectNameX, ...)

\method{extractTags}{htest}(object, objectNameX, ...)

\method{extractTags}{lda}(object, objectNameX, ...)

\method{extractTags}{qda}(object, objectNameX, ...)

\method{extractTags}{twins}(object, objectNameX, ...)

\method{extractTags}{partition}(object, objectNameX, ...)

\method{extractTags}{glmnet}(object, objectNameX, ...)

\method{extractTags}{survfit}(object, objectNameX, ...)
}
\arguments{
\item{object}{for this object tags are to be}

\item{objectNameX}{name of the object}

\item{...}{other arguments}
}
\value{
\code{NULL}

\code{NULL}

\code{NULL}

\code{NULL}

\code{NULL}

\code{NULL}

\code{NULL}

\code{NULL}

\code{NULL}

\code{NULL}

\code{NULL}

\code{NULL}

\code{NULL}
}
\description{
Internal function for extraction of tags from objects
}
