% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRemoteHook.R
\name{getRemoteHookGithub}
\alias{getRemoteHookGithub}
\title{Get http Hook for Remote Repo}
\usage{
getRemoteHookGithub(repo = NULL, user = NULL, branch = "master",
  repoDirGit = FALSE)
}
\arguments{
\item{repo}{A character containing a name of a Git repository on which the Repository is archived.}

\item{user}{A character containing a name of a Git user on whose account the \code{repo} is created.}

\item{branch}{A character containing a name of Git Repository's branch on which the Repository is archived. 
Default \code{branch} is \code{master}.}

\item{repoDirGit}{A character containing a name of a directory on Git repository 
on which the Repository is stored. If the Repository is stored in main folder on Git repository, this should be set 
to \code{repoDirGit = FALSE} as default.}
}
\description{
\code{loadFromLocalRepo} loads an artifact from a local \link{Repository} into the workspace.
\code{loadFromGithubRepo} loads an artifact from a Github \link{Repository} into the workspace.
\code{loadFromRemoteRepo} loads an artifact from a git / mercurial \link{Repository} into the workspace.
\code{loadFromRepo} is a wrapper around \code{loadFromLocalRepo} and \code{loadFromGithubRepo}.
To learn more about artifacts visit \link[archivist]{archivist-package}.
}
\examples{

\dontrun{
# objects preparation
}
}
\author{
Przemysla Biecek, \email{przemyslaw.biecek@gmail.com}
}
\seealso{
Other archivist: \code{\link{Repository}},
  \code{\link{Tags}}, \code{\link{\%a\%}},
  \code{\link{addTagsRepo}}, \code{\link{ahistory}},
  \code{\link{alink}}, \code{\link{aoptions}},
  \code{\link{archive}},
  \code{\link{archivist-github-integration}},
  \code{\link{archivist-package}}, \code{\link{aread}},
  \code{\link{asearch}}, \code{\link{cache}},
  \code{\link{cloneGithubRepo}},
  \code{\link{copyLocalRepo}},
  \code{\link{createEmptyRepo}},
  \code{\link{createGithubMDGallery}},
  \code{\link{deleteRepo}}, \code{\link{getTagsLocal}},
  \code{\link{loadFromLocalRepo}}, \code{\link{md5hash}},
  \code{\link{pushRepo}}, \code{\link{rmFromRepo}},
  \code{\link{saveToRepo}},
  \code{\link{searchInLocalRepo}},
  \code{\link{setLocalRepo}},
  \code{\link{shinySearchInLocalRepo}},
  \code{\link{showLocalRepo}},
  \code{\link{splitTagsLocal}},
  \code{\link{summaryLocalRepo}},
  \code{\link{zipLocalRepo}}
}

