% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pushRepo.R
\name{pushRepo}
\alias{pushRepo}
\title{Push Local Repository to GitHub}
\usage{
pushRepo(repoDir = aoptions("repoDir"),
  commitMessage = aoptions("commitMessage"), repo = aoptions("repo"),
  user = aoptions("user"), password = aoptions("password"),
  files = c("gallery", "backpack.db"))
}
\arguments{
\item{repoDir}{A character specifing the directory to Local \code{Repository} from which artifacts will be pushed to GitHub.}

\item{commitMessage}{A character denoting a message added to the commit while performing push.
By default specified to \code{NULL} which corresponds to commit message \code{archivist: pushRepo}.}

\item{repo}{A character denoting GitHub repository name and synchronized local existing directory in which an artifact will be saved.}

\item{user}{A character denoting GitHub user name. Can be set globally with \code{aoptions("user", user)}.
See \link{archivist-github-integration}.}

\item{password}{A character denoting GitHub user password. Can be set globally with \code{aoptions("password", password)}.
See \link{archivist-github-integration}.}

\item{files}{A character vector containing directories to files that should be commited and pushed. The working directory
is \code{repoDir}. By default all uncommited artifacts and \code{backpack.db} will be pushed.}
}
\description{
\code{pushRepo} add files, commits them and pushes from Local \link{Repository} to synchronized GitHub one.
}
\examples{
\dontrun{

# TODO

}
}
\author{
Marcin Kosinski, \email{m.p.kosinski@gmail.com}
}
\seealso{
Other archivist: \code{\link{Repository}},
  \code{\link{Tags}}, \code{\link{\%a\%}},
  \code{\link{addTagsRepo}}, \code{\link{ahistory}},
  \code{\link{alink}}, \code{\link{aoptions}},
  \code{\link{archive}},
  \code{\link{archivist-github-integration}},
  \code{\link{archivist-package}}, \code{\link{aread}},
  \code{\link{asearch}}, \code{\link{cache}},
  \code{\link{cloneGithubRepo}},
  \code{\link{copyLocalRepo}},
  \code{\link{createEmptyRepo}},
  \code{\link{createGithubMDGallery}},
  \code{\link{deleteRepo}}, \code{\link{getTagsLocal}},
  \code{\link{loadFromLocalRepo}}, \code{\link{md5hash}},
  \code{\link{rmFromRepo}}, \code{\link{saveToRepo}},
  \code{\link{searchInLocalRepo}},
  \code{\link{setLocalRepo}},
  \code{\link{shinySearchInLocalRepo}},
  \code{\link{showLocalRepo}},
  \code{\link{splitTagsLocal}},
  \code{\link{summaryLocalRepo}},
  \code{\link{zipLocalRepo}}
}

