% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restoreLibraries.R
\name{restoreLibs}
\alias{restoreLibs}
\title{Restore Versions of Libraries}
\usage{
restoreLibs(md5hash, session_info = NULL)
}
\arguments{
\item{md5hash}{One of the following (see \link{aread}):

A character vector which elements  are consisting of at least three components separated with '/': Remote user name, Remote repository and name of the artifact (MD5 hash) or it's abbreviation.

MD5 hashes of artifacts in current local default directory or its abbreviations.}
}
\value{
This function returns a list of artifacts (by their values).
}
\description{
Function \code{restoreLibrary} gets either session_info or artifact's id 
and restore libraries to the version set in the object.
}
\examples{
\dontrun{
## objects preparation
restoreLibs(md5hash = "pbiecek/graphGallery/600bda83cb840947976bd1ce3a11879d")
}
}
\author{
Marcin Kosinski, \email{m.p.kosinski@gmail.com}
Przemyslaw Biecek, \email{przemyslaw.biecek@gmail.com}
}
\seealso{
Other archivist: \code{\link{Repository}},
  \code{\link{Tags}}, \code{\link{\%a\%}},
  \code{\link{addHooksToPrint}}, \code{\link{addTagsRepo}},
  \code{\link{aformat}}, \code{\link{ahistory}},
  \code{\link{alink}}, \code{\link{aoptions}},
  \code{\link{archivist-package}}, \code{\link{aread}},
  \code{\link{asearch}}, \code{\link{asession}},
  \code{\link{cache}}, \code{\link{copyLocalRepo}},
  \code{\link{createLocalRepo}},
  \code{\link{createMDGallery}},
  \code{\link{deleteLocalRepo}},
  \code{\link{getRemoteHook}}, \code{\link{getTagsLocal}},
  \code{\link{loadFromLocalRepo}}, \code{\link{md5hash}},
  \code{\link{rmFromLocalRepo}},
  \code{\link{saveToLocalRepo}},
  \code{\link{searchInLocalRepo}},
  \code{\link{setLocalRepo}},
  \code{\link{shinySearchInLocalRepo}},
  \code{\link{showLocalRepo}},
  \code{\link{splitTagsLocal}},
  \code{\link{summaryLocalRepo}},
  \code{\link{zipLocalRepo}}
}

