% Generated by roxygen2 (4.0.1): do not edit by hand
\name{returnTagLocal}
\alias{returnTagGithub}
\alias{returnTagLocal}
\title{Return a \code{Tag} Corresponding to \code{md5hash}}
\usage{
returnTagLocal(md5hash, repoDir, tag = "name")

returnTagGithub(md5hash, user, repo, branch = "master", tag = "name")
}
\arguments{
\item{repoDir}{A character denoting an existing directory in
which an artifact is stored.}

\item{md5hash}{A character containing \code{md5hash}
of artifacts which \code{Tag} is desired to be returned.}

\item{tag}{A type of a \link{Tags}. Default \code{tag = "name"}.}

\item{repo}{Only if working with a Github repository. A character containing a name of a Github repository on which the Repository is archived.}

\item{user}{Only if working with a Github repository. A character containing a name of a Github user on whose account the \code{repo} is created.}

\item{branch}{Only if working with a Github repository. A character containing a name of
Github repository's branch in which Repository is archived. Default \code{branch} is \code{master}.}
}
\value{
The returned value is a \code{Tag} (see \link{Tags}) related to \link{md5hash}
of an artifact, which is a character.
}
\description{
\code{returnTagLocal} and \code{returnTagGithub} return a \code{Tag} (see \link{Tags}) related to \link{md5hash}
of an artifact. To learn more about artifacts visit \link[archivist]{archivist-package}.
}
\details{
\code{returnTagLocal} and \code{returnTagGithub} return (see \link{Tags}) related to \link{md5hash}
of an artifact. To learn more about artifacts visit \link[archivist]{archivist-package}.
}
\examples{
\dontrun{
library(dplyr)
exampleRepoDir <- tempdir()
createEmptyRepo( exampleRepoDir )

data(mtcars)
hash <- mtcars \%.\%
 group_by(cyl, am) \%.\%
 select(mpg, cyl, wt, am) \%.\%
 summarise(avgmpg = mean(mpg), avgwt = mean(wt)) \%.\%
 filter(avgmpg > 20) \%>\%
 saveToRepo( exampleRepoDir )

returnTagLocal( md5hash = hash, exampleRepoDir )

deleteRepo( exampleRepoDir )
rm( exampleRepoDir )

returnTagGithub( "3db63bc63b8defaf42c0bde19160f242",
   user="pbiecek", repo="archivist")

}
}
\author{
Marcin Kosinski, \email{m.p.kosinski@gmail.com}
}
\seealso{
Other archivist: \code{\link{Repository}};
  \code{\link{Tags}}; \code{\link{archivist-package}};
  \code{\link{copyGithubRepo}},
  \code{\link{copyLocalRepo}};
  \code{\link{createEmptyRepo}}; \code{\link{deleteRepo}};
  \code{\link{loadFromGithubRepo}},
  \code{\link{loadFromLocalRepo}}; \code{\link{md5hash}};
  \code{\link{rmFromRepo}}; \code{\link{saveToRepo}};
  \code{\link{searchInGithubRepo}},
  \code{\link{searchInLocalRepo}};
  \code{\link{showGithubRepo}},
  \code{\link{showLocalRepo}};
  \code{\link{summaryGithubRepo}},
  \code{\link{summaryLocalRepo}}

Other archivist: \code{\link{Repository}};
  \code{\link{Tags}}; \code{\link{archivist-package}};
  \code{\link{copyGithubRepo}},
  \code{\link{copyLocalRepo}};
  \code{\link{createEmptyRepo}}; \code{\link{deleteRepo}};
  \code{\link{loadFromGithubRepo}},
  \code{\link{loadFromLocalRepo}}; \code{\link{md5hash}};
  \code{\link{rmFromRepo}}; \code{\link{saveToRepo}};
  \code{\link{searchInGithubRepo}},
  \code{\link{searchInLocalRepo}};
  \code{\link{showGithubRepo}},
  \code{\link{showLocalRepo}};
  \code{\link{summaryGithubRepo}},
  \code{\link{summaryLocalRepo}}
}

