% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rmFromRepo}
\alias{rmFromRepo}
\title{Remove an Object from a Repository}
\usage{
rmFromRepo(md5hash, dir)
}
\arguments{
\item{md5hash}{A hash of an object. A character string being a result of a cryptographical hash function with MD5 algorithm.}

\item{dir}{A character denoting an existing directory from which an object will be removed.}
}
\description{
\code{rmFromRepo} removes an object given as \code{md5hash} from a Repository.
}
\details{
\code{rmFromRepo} removes an object given as \code{md5hash} from a Repository,
which is a SQLite database named \code{backpack} - created by a \link{createEmptyRepo} call.
\code{md5hash} is a string of length 32 that comes out as a result from \link{saveToRepo} function,
which uses a cryptographical hash function with MD5 algorithm.

Also this function removes a \code{md5hash.rda} file, where \code{md5hash} is object's hash as above.

Important: instead of giving whole \code{md5hash} name, user can simply give first few signs of desired \code{md5hash} - an abbreviation.
For example \code{a09dd} instead of \code{a09ddjdkf9kj33dcjdnfjgos9jd9jkcv}. But if several tags start with the same pattern
an error will be displayed and you will be asked to give more precise \code{md5hash} abbreviation (try abbreviation with more digits all with whole name).
}
\note{
\code{md5hash} can be possessed as a value of a \link{searchInRepo} function proceeded with \code{tag = NAME} argument,
where \code{NAME} is desired to remove object's name.

For more informations about \code{Tags} check \link{Tags}.
}
\examples{
# not work
rmFromRepo( md5hash = "sksjdncjdkslwifgdtsjhdt37wyshzdh",
                       dir = "REPODIR" )

# with md5hash abbreviation
rmFromRepo( md5hash = "sksjd",
                       dir = "REPODIR" )

 hash <-  searchInLocalRepo( "name:prettyPlot",
                       dir = "/home/folder/here" )
 rmFromRepo( hash, dir = "/home/folder/here" )
}
\author{
Marcin Kosinski , \email{m.p.kosinski@gmail.com}
}
\seealso{
Other archivist: \code{\link{Tags}};
  \code{\link{archivist-package}};
  \code{\link{createEmptyRepo}};
  \code{\link{loadFromGithubRepo}},
  \code{\link{loadFromLocalRepo}};
  \code{\link{saveToRepo}};
  \code{\link{searchInGithubRepo}},
  \code{\link{searchInLocalRepo}}
}

