% Generated by roxygen2 (4.0.1): do not edit by hand
\name{saveToRepo}
\alias{saveToRepo}
\title{Save an Object into a Repository}
\usage{
saveToRepo(object, ..., archiveData = TRUE, archiveTags = TRUE,
  archiveMiniature = TRUE, dir)
}
\arguments{
\item{object}{An arbitrary R object to be saved.}

\item{...}{Graphical parameters denoting width and hight of a miniature. See details.}

\item{archiveData}{A logical value denoting wheter to archive the data from object.}

\item{archiveTags}{A logical value denoting wheter to archive tags from object.}

\item{archiveMiniature}{A logical value denoting wheter to archive a miniature of an object.}

\item{dir}{A character denoting an existing directory in which an object will be saved.}
}
\description{
desc
}
\details{
details
}
\examples{
#example
}
\author{
autor
}
\seealso{
Other archivist: \code{\link{createEmptyRepo}};
  \code{\link{loadFromGithubRepo}},
  \code{\link{loadFromLocalRepo}};
  \code{\link{rmFromRepo}};
  \code{\link{searchInGithubRepo}},
  \code{\link{searchInLocalRepo}}
}

