% Generated by roxygen2 (4.0.1): do not edit by hand
\name{searchInLocalRepo}
\alias{searchInGithubRepo}
\alias{searchInLocalRepo}
\title{Search for an Object in a Repository using Tags}
\usage{
searchInLocalRepo(tag, dir)

searchInGithubRepo(tag, repo, user)
}
\arguments{
\item{tag}{A character string denoting a Tag to seek for. See details.}

\item{dir}{A character denoting an existing directory from which objects will be searched.}

\item{repo}{Only if working on Github Repository. A character string containing a name of Github Repository.}

\item{user}{Only if working on Github Repository. A character string containing a name of Github User.}
}
\description{
desc
}
\details{
details
}
\examples{
#example
}
\author{
autor
}
\seealso{
Other archivist: \code{\link{createEmptyRepo}};
  \code{\link{loadFromGithubRepo}},
  \code{\link{loadFromLocalRepo}};
  \code{\link{rmFromRepo}}; \code{\link{saveToRepo}}
}

