% Generated by roxygen2 (4.0.1): do not edit by hand
\name{setLocalRepo}
\alias{setGlobalRepo}
\alias{setLocalRepo}
\title{Set Global Path to Repository}
\usage{
setLocalRepo(repoDir)

setGlobalRepo(user, repo, branch = "master", repoDirGit = NULL)
}
\arguments{
  \item{repoDir}{}

  \item{repo}{Only if working with a Github repository. A
  character containing a name of a Github repository on
  which the Repository is archived.}

  \item{user}{Only if working with a Github repository. A
  character containing a name of a Github user on whose
  account the \code{repo} is created.}

  \item{branch}{Only if working with a Github repository. A
  character containing a name of Github Repository's branch
  on which the Repository is archived. Default
  \code{branch} is \code{master}.}

  \item{repoDirGit}{Only if working with a Github
  repository. A character containing a name of a directory
  on Github repository on which the Repository is stored.
  If the Repository is stored in main folder on Github
  repository, this should be set to \code{repoDirGit =
  FALSE} as default.}
}
\description{
\code{setLocalRepo} \code{setGlobalRepo} To learn more
about artifacts visit \link[archivist]{archivist-package}.
}
\details{
Functions
}
\examples{
\dontrun{
# examples
exampleRepoDir <- tempdir()
createEmptyRepo( repoDir = exampleRepoDir )
saveToRepo( iris, exampleRepoDir)
setLocalRepo( exampleRepoDir )
saveToRepo( swiss )
showLocalRepo( exampleRepoDir )

iris2 <- loadFromLocalRepo( "ff575c2" , value = TRUE)
head(iris2, 2)

searchInLocalRepo( "name:i", fixed = F)

getTagsLocal( "ff575c261c949d073b2895b05d1097c3" )

rmFromRepo( "4c43f" )
showLocalRepo( )
summaryLocalRepo( )

deleteRepo( exampleRepoDir )
rm( exampleRepoDir )

}
}
\author{
Marcin Kosinski , \email{m.p.kosinski@gmail.com}
}
\seealso{
\href{https://github.com/pbiecek/archivist/wiki}{https://github.com/pbiecek/archivist/wiki}

Other archivist: \code{\link{Repository}};
\code{\link{Tags}}; \code{\link{addTagsRepo}};
\code{\link{archivist-package}}; \code{\link{cache}};
\code{\link{copyGithubRepo}}, \code{\link{copyLocalRepo}};
\code{\link{createEmptyRepo}}; \code{\link{deleteRepo}};
\code{\link{getTagsGithub}}, \code{\link{getTagsLocal}};
\code{\link{loadFromGithubRepo}},
\code{\link{loadFromLocalRepo}}; \code{\link{md5hash}};
\code{\link{multiSearchInGithubRepo}},
\code{\link{multiSearchInLocalRepo}},
\code{\link{searchInGithubRepo}},
\code{\link{searchInLocalRepo}}; \code{\link{rmFromRepo}};
\code{\link{saveToRepo}};
\code{\link{shinySearchInLocalRepo}};
\code{\link{showGithubRepo}}, \code{\link{showLocalRepo}};
\code{\link{summaryGithubRepo}},
\code{\link{summaryLocalRepo}};
\code{\link{zipGithubRepo}}, \code{\link{zipLocalRepo}}

Other archivist: \code{\link{Repository}};
\code{\link{Tags}}; \code{\link{addTagsRepo}};
\code{\link{archivist-package}}; \code{\link{cache}};
\code{\link{copyGithubRepo}}, \code{\link{copyLocalRepo}};
\code{\link{createEmptyRepo}}; \code{\link{deleteRepo}};
\code{\link{getTagsGithub}}, \code{\link{getTagsLocal}};
\code{\link{loadFromGithubRepo}},
\code{\link{loadFromLocalRepo}}; \code{\link{md5hash}};
\code{\link{multiSearchInGithubRepo}},
\code{\link{multiSearchInLocalRepo}},
\code{\link{searchInGithubRepo}},
\code{\link{searchInLocalRepo}}; \code{\link{rmFromRepo}};
\code{\link{saveToRepo}};
\code{\link{shinySearchInLocalRepo}};
\code{\link{showGithubRepo}}, \code{\link{showLocalRepo}};
\code{\link{summaryGithubRepo}},
\code{\link{summaryLocalRepo}};
\code{\link{zipGithubRepo}}, \code{\link{zipLocalRepo}}
}

