% Generated by roxygen2 (4.0.2): do not edit by hand
\name{shinySeachInLocalRepo}
\alias{shinySeachInLocalRepo}
\title{Shiny Based Live Search for an Artifact in a Repository Using Tags}
\usage{
shinySeachInLocalRepo(repoDir, host = "0.0.0.0")
}
\arguments{
  \item{repoDir}{A character denoting an existing directory
  in which artifacts will be searched.}

  \item{host}{A host IP adress, see the \code{host}
  argument in \code{shiny::runApp}.}
}
\value{
\code{shinySeachInLocalRepo} runs a shiny application.
}
\description{
\code{shinySeachInLocalRepo} searches for an artifact in a
\link{Repository} using it's \link{Tags}. An shiny base
application is generated in the fly with an text input
field. To learn more about artifacts visit
\link[archivist]{archivist-package}.
}
\details{
\code{shinySeachInLocalRepo} searches for an artifact in a
Repository using it's \code{Tag} (e.g., \code{name},
\code{class} or \code{archiving date}). \code{Tags} are
submitted in a text input in a shiny application. Many tags
may be specified, they should be comma separated.

\code{Tags}, submitted in the text field, should be given
according to the format: \code{"TagType:TagTypeValue"} -
see examples.
}
\examples{
\dontrun{
  # assuming that there is a 'repo' dir with a valid archivist repository
  shinySeachInLocalRepo( dir = 'repo' )
}
}
\author{
Przemyslaw Biecek, \email{przemyslaw.biecek@gmail.com}
}
\seealso{
Other archivist: \code{\link{Repository}};
\code{\link{Tags}}; \code{\link{archivist-package}};
\code{\link{copyGithubRepo}}, \code{\link{copyLocalRepo}};
\code{\link{createEmptyRepo}}; \code{\link{deleteRepo}};
\code{\link{getTagsGithub}}, \code{\link{getTagsLocal}};
\code{\link{loadFromGithubRepo}},
\code{\link{loadFromLocalRepo}}; \code{\link{md5hash}};
\code{\link{multiSearchInGithubRepo}},
\code{\link{multiSearchInLocalRepo}},
\code{\link{searchInGithubRepo}},
\code{\link{searchInLocalRepo}}; \code{\link{rmFromRepo}};
\code{\link{saveToRepo}}; \code{\link{showGithubRepo}},
\code{\link{showLocalRepo}};
\code{\link{summaryGithubRepo}},
\code{\link{summaryLocalRepo}};
\code{\link{tarGithubRepo}}, \code{\link{tarLocalRepo}}
}

