% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showRepo.R
\name{showLocalRepo}
\alias{showGithubRepo}
\alias{showLocalRepo}
\title{View the List of Artifacts from the Repository}
\usage{
showLocalRepo(repoDir = NULL, method = "md5hashes")

showGithubRepo(repo = NULL, user = NULL, branch = "master",
  repoDirGit = FALSE, method = "md5hashes")
}
\arguments{
\item{repoDir}{A character denoting an existing directory of the Repository for which metadata will be returned.
If it is set to \code{NULL} (by default), it will use the \code{repoDir} specified in \link{setLocalRepo}.}

\item{method}{A character specifying a method to be used to show the Repository. Available methods: 
\code{md5hashes} (default), \code{tags} and \code{sets} - see \href{https://github.com/pbiecek/archivist2}{archivist2::saveSetToRepo}.}

\item{repo}{While working with the Github repository. A character containing a name of the Github repository on which the Repository is stored.
By default set to \code{NULL} - see \code{Note}.}

\item{user}{While working with the Github repository. A character containing a name of the Github user on whose account the \code{repo} is created.
By default set to \code{NULL} - see \code{Note}.}

\item{branch}{While working with the Github repository. A character containing a name of 
the Github Repository's branch on which the Repository is stored. Default \code{branch} is \code{master}.}

\item{repoDirGit}{While working with the Github repository. A character containing a name of a directory on the Github repository 
on which the Repository is stored. If the Repository is stored in the main folder of the Github repository, this should be set 
to \code{repoDirGit = FALSE} as default.}
}
\value{
If parameter \code{method} is set as \code{md5hashes} then a \code{data.frame} with artifacts' names and
artifacts'\code{md5hashes} will be returned.

If parameter \code{method} is set as \code{tags} then a \code{data.frame} with \code{Tags} and
artifacts' \code{md5hashes} will be returned.

Also in both cases a \code{data.frame} contains an extra column with the date of creation
of the \code{Tag} or \code{md5hash}.

To learn more about \code{Tags} or \code{md5hashes} check: \link{Tags} or \link{md5hash}.
}
\description{
\code{showLocalRepo} and \code{showGithubRepo} functions produce the \code{data.frame} of the artifacts from
the \link{Repository} saved in a given \code{repoDir} (directory). \code{showLocalRepo}
shows the artifacts from the \code{Repository} that exists on the user's computer whereas \code{showGithubRepo}
shows the artifacts of the \code{Repository} existing on the Github repository.
To learn more about artifacts visit \link[archivist]{archivist-package}.
}
\details{
\code{showLocalRepo} and \code{showGithubRepo} functions produce the \code{data.frame} of the artifacts from
a \link{Repository} saved in a given \code{repoDir} (directory). \code{showLocalRepo}
shows the artifacts from the \code{Repository} that exists on the user's computer whereas \code{showGithubRepo}
shows the artifacts of the \code{Repository} existing on the Github repository.

Both functions show the current state of a \code{Repository}, inter alia, all archived artifacts can
be seen with their unique \link{md5hash} or a \code{data.frame} with archived \link{Tags} can 
be obtained.
}
\note{
If \code{repo} and \code{user} are set to \code{NULL} (as default) in the Github mode then global parameters
set in \link{setGithubRepo} function are used.
}
\examples{
\dontrun{
# objects preparation

# data.frame object
data(iris)

# ggplot/gg object
library(ggplot2)
df <- data.frame(gp = factor(rep(letters[1:3], each = 10)),y = rnorm(30))
library(plyr)
ds <- ddply(df, .(gp), summarise, mean = mean(y), sd = sd(y))
myplot123 <- ggplot(df, aes(x = gp, y = y)) +
  geom_point() +  geom_point(data = ds, aes(y = mean),
               colour = 'red', size = 3)
               
# lm object                
model <- lm(Sepal.Length~ Sepal.Width + Petal.Length + Petal.Width, data= iris)


# lda object
library(MASS)

Iris <- data.frame(rbind(iris3[,,1], iris3[,,2], iris3[,,3]),
                  Sp = rep(c("s","c","v"), rep(50,3)))
train <- c(8,83,115,118,146,82,76,9,70,139,85,59,78,143,68,
           134,148,12,141,101,144,114,41,95,61,128,2,42,37,
           29,77,20,44,98,74,32,27,11,49,52,111,55,48,33,38,
           113,126,24,104,3,66,81,31,39,26,123,18,108,73,50,
           56,54,65,135,84,112,131,60,102,14,120,117,53,138,5)
lda1 <- lda(Sp ~ ., Iris, prior = c(1,1,1)/3, subset = train)

# qda object
tr <- c(7,38,47,43,20,37,44,22,46,49,50,19,4,32,12,29,27,34,2,1,17,13,3,35,36)
train <- rbind(iris3[tr,,1], iris3[tr,,2], iris3[tr,,3])
cl <- factor(c(rep("s",25), rep("c",25), rep("v",25)))
qda1 <- qda(train, cl)

# glmnet object
library( glmnet )

zk=matrix(rnorm(100*20),100,20)
bk=rnorm(100)
glmnet1=glmnet(zk,bk)

# creating example Repository - on which examples will work

exampleRepoDir <- tempfile()
createEmptyRepo(repoDir = exampleRepoDir)
saveToRepo(myplot123, repoDir=exampleRepoDir)
saveToRepo(iris, repoDir=exampleRepoDir)
saveToRepo(model, repoDir=exampleRepoDir)

# show examples

showLocalRepo(method = "md5hashes", repoDir = exampleRepoDir)
showLocalRepo(method = "tags", repoDir = exampleRepoDir)

# let's add more artifacts

saveToRepo(glmnet1, repoDir=exampleRepoDir)
saveToRepo(lda1, repoDir=exampleRepoDir)
(qda1Md5hash <- saveToRepo(qda1, repoDir=exampleRepoDir))

# show now

showLocalRepo(method = "md5hashes", repoDir = exampleRepoDir)
showLocalRepo(method = "tags", repoDir = exampleRepoDir)

# what if we remove an artifact

rmFromRepo(qda1Md5hash, repoDir = exampleRepoDir)

# show now

showLocalRepo(method = "md5hashes", repoDir = exampleRepoDir)
showLocalRepo(method = "tags", repoDir = exampleRepoDir)

# GitHub version

showGithubRepo(method = "md5hashes", user = "pbiecek", repo = "archivist")
showGithubRepo(method = "tags", user = "pbiecek", repo = "archivist", branch = "master")



# sets method
setLocalRepo( exampleRepoDir )
library(ggplot2)
library(ggthemes)

data(iris)

plotArtifact <- ggplot( iris, aes(x = Sepal.Length, y = Species)) +
  geom_point()+
  theme_wsj()

plotData <- iris
plotFunctions <- list( ggplot, geom_point, theme_wsj)

# removing an example Repository
  
deleteRepo( exampleRepoDir, TRUE )
  
rm( exampleRepoDir )

# many archivist-like Repositories on one Github repository

showGithubRepo( user="MarcinKosinski", repo="Museum", branch="master",
repoDirGit="ex1")
showGithubRepo( user="MarcinKosinski", repo="Museum", branch="master",
                repoDirGit="ex2")
                
## Github options
showGithubRepo('archivist', 'pbiecek')
aoptions('user', 'pbiecek')
aoptions('repo', 'archivist')
loadFromGithubRepo("ff575c261c", value = TRUE) -> iris123

showGithubRepo('Museum', 'MarcinKosinski', repoDirGit = 'ex1')
aoptions('repo', 'Museum')
aoptions('user', 'MarcinKosinski')
aoptions('repoDirGit', 'ex1')
aoptions('branch', 'master')
showGithubRepo()
showGithubRepo(repoDirGit = 'ex2')

aoptions('repoDirGit')


}
}
\author{
Marcin Kosinski , \email{m.p.kosinski@gmail.com}
}
\seealso{
Other archivist: \code{\link{Repository}},
  \code{\link{Tags}}, \code{\link{\%a\%}},
  \code{\link{addTagsRepo}}, \code{\link{ahistory}},
  \code{\link{alink}}, \code{\link{aoptions}},
  \code{\link{archive}},
  \code{\link{archivist-github-integration}},
  \code{\link{archivist-package}}, \code{\link{aread}},
  \code{\link{asearch}}, \code{\link{cache}},
  \code{\link{cloneGithubRepo}},
  \code{\link{copyLocalRepo}},
  \code{\link{createEmptyRepo}}, \code{\link{deleteRepo}},
  \code{\link{getTagsLocal}},
  \code{\link{loadFromLocalRepo}}, \code{\link{md5hash}},
  \code{\link{rmFromRepo}}, \code{\link{saveToRepo}},
  \code{\link{searchInLocalRepo}},
  \code{\link{setLocalRepo}},
  \code{\link{shinySearchInLocalRepo}},
  \code{\link{splitTagsLocal}},
  \code{\link{summaryLocalRepo}},
  \code{\link{zipLocalRepo}}
}

