% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{Tags}
\alias{Tags}
\title{Tags}
\description{
\code{Tags} are specific values of an object, different for various object's classes.
}
\details{
\code{Tags} are specific values of an object. \code{Tags} can be an object's \code{name}, \code{class} or \code{archivisation date}.
Furthermore, for various object's classes more different \code{Tags} are available and can
be searched by \link{searchInLocalRepo} or \link{searchInGithubRepo} functions. Supported objects with
extra \code{Tags} are:
}
\seealso{
Functions using \code{Tags} are \link{searchInLocalRepo} and \link{searchInGithubRepo}.

Other archivist: \code{\link{archivist-package}};
  \code{\link{createEmptyRepo}};
  \code{\link{loadFromGithubRepo}},
  \code{\link{loadFromLocalRepo}};
  \code{\link{rmFromRepo}}; \code{\link{saveToRepo}};
  \code{\link{searchInGithubRepo}},
  \code{\link{searchInLocalRepo}}
}

