% Generated by roxygen2 (4.0.1): do not edit by hand
\name{tarLocalRepo}
\alias{tarLocalRepo}
\title{Create a Tar Archive From an Existing Repository}
\usage{
tarLocalRepo(repoDir)
}
\arguments{
\item{repoDir}{A character that specifies the directory of the Repository which
will be tarred.}
}
\description{
\code{tarLocalRepo} ...
}
\details{
\code{tarLocalRepo}
}
\examples{
\dontrun{

}
}
\author{
Marcin Kosinski, \email{m.p.kosinski@gmail.com}
}
\seealso{
Other archivist: \code{\link{Repository}};
  \code{\link{Tags}}; \code{\link{archivist-package}};
  \code{\link{copyGithubRepo}},
  \code{\link{copyLocalRepo}};
  \code{\link{createEmptyRepo}}; \code{\link{deleteRepo}};
  \code{\link{getTagsGithub}}, \code{\link{getTagsLocal}};
  \code{\link{loadFromGithubRepo}},
  \code{\link{loadFromLocalRepo}}; \code{\link{md5hash}};
  \code{\link{rmFromRepo}}; \code{\link{saveToRepo}};
  \code{\link{searchInGithubRepo}},
  \code{\link{searchInLocalRepo}};
  \code{\link{showGithubRepo}},
  \code{\link{showLocalRepo}};
  \code{\link{summaryGithubRepo}},
  \code{\link{summaryLocalRepo}}
}

