% Generated by roxygen2 (4.0.1): do not edit by hand
\name{addTagsRepo}
\alias{addTagsRepo}
\title{Add new Tags to the Existing Repository}
\usage{
addTagsRepo(x)
}
\arguments{
  \item{md5hashes}{To which corresponding artifacts should
  \code{Tags} be added.}

  \item{tags}{A character vector that specifies what tag
  should be added to which artifact that corresponds to
  given \code{md5hash}. Can be specified only one of:
  \code{FUN} or \code{tags}. Should have lenght 1 or the
  same as lenght of \code{md5hashes} param.}

  \item{repoDir}{A character that specifies the directory
  of the Repository into which new \code{Tags} will be
  added.}

  \item{FUN}{A function that evaluates on the artifacts for
  which \code{md5hashes} are given and generates
  \code{Tags} as a result that will be added to the local
  Repository. Can be specified only one of: \code{FUN} or
  \code{tags}.}
}
\description{
\code{addTagsRepo}
}
\details{
\describe{ \item{Problem:}{ \itemize{ \item after an
artifact is added there is no way to add/update it's tags
sometimes one discover, that it would be nice to have
additional properties exposed as tags } } \item{Proposed
solution:}{ \itemize{ \item addTag(md5hashes, repoDir, FUN,
tags) function, that will take list of md5hashes if user
specifies tags, then these tags will be added to all
md5hashes if user specifies FUN, then FUN will be executed
on each md5hash object, FUN returns a character vector
(=new tags), new tags are added to database } } }
}
\examples{
\dontrun{

# Takes all objects of lm class from repository,
# extracts R2 for them and stores as R2: tags

md5hashes <- searchInLocalRepo(repoDir, “class:lm”)
addTag(md5hashes, repoDir, function(x) paste0(“R2:”,summary(x)$r.square))
}
}
\author{
Marcin Kosinski, \email{m.p.kosinski@gmail.com}
}
\seealso{
Other archivist: \code{\link{Repository}};
\code{\link{Tags}}; \code{\link{archivist-package}};
\code{\link{cache}}; \code{\link{copyGithubRepo}},
\code{\link{copyLocalRepo}}; \code{\link{createEmptyRepo}};
\code{\link{deleteRepo}}; \code{\link{getTagsGithub}},
\code{\link{getTagsLocal}};
\code{\link{loadFromGithubRepo}},
\code{\link{loadFromLocalRepo}}; \code{\link{md5hash}};
\code{\link{multiSearchInGithubRepo}},
\code{\link{multiSearchInLocalRepo}},
\code{\link{searchInGithubRepo}},
\code{\link{searchInLocalRepo}}; \code{\link{rmFromRepo}};
\code{\link{saveToRepo}};
\code{\link{shinySearchInLocalRepo}};
\code{\link{showGithubRepo}}, \code{\link{showLocalRepo}};
\code{\link{summaryGithubRepo}},
\code{\link{summaryLocalRepo}};
\code{\link{zipGithubRepo}}, \code{\link{zipLocalRepo}}
}

