% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{archivist-package}
\alias{archivist-package}
\title{Tools for archivisation of R objects}
\description{
TODO: A description.
}
\details{
TODO: To be written.
}
\author{
Przemyslaw Biecek [aut, cre] \email{przemyslaw.biecek@gmail.com} \cr
Marcin Kosinski [aut] \email{m.p.kosinski@gmail.com}
}
\seealso{
Other archivist: \code{\link{Repository}};
  \code{\link{Tags}}; \code{\link{createEmptyRepo}};
  \code{\link{loadFromGithubRepo}},
  \code{\link{loadFromLocalRepo}}; \code{\link{md5hash}};
  \code{\link{rmFromRepo}}; \code{\link{saveToRepo}};
  \code{\link{searchInGithubRepo}},
  \code{\link{searchInLocalRepo}};
  \code{\link{summaryGithubRepo}},
  \code{\link{summaryLocalRepo}}
}

