% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{archivist-package}
\alias{archivist-package}
\title{Tools for archivisation of R objects}
\description{
Data exploration and modelling is a process in which a lot of data artifacts are produced.
Artifacts like: subsets, data aggregates, plots, statistical models, different
versions of data sets and different versions of results.
The more projects we work with the more artifacts are produced and the harder it is to manage these artifacts.

Archivist helps to store and manage artifacts created in R.

Archivist allows you to store selected artifacts as a binary files
  together with their metadata and relations.
Archivist allows to share artifacts with others, either through shared folder or github.
Archivist allows to look for already created artifacts by using it's class,
  name, date of the creation or other properties.
Makes it easy to restore such artifacts.
Archivist allows to check if new artifact is the exact
  copy that was produced some time ago.
  That might be usefull either for testing or caching.
}
\details{
TODO: To be written.
}
\author{
Przemyslaw Biecek [aut, cre] \email{przemyslaw.biecek@gmail.com} \cr
Marcin Kosinski [aut] \email{m.p.kosinski@gmail.com}
}
\seealso{
Other archivist: \code{\link{Repository}};
  \code{\link{Tags}}; \code{\link{createEmptyRepo}};
  \code{\link{deleteRepo}}; \code{\link{infoGithubRepo}},
  \code{\link{infoLocalRepo}};
  \code{\link{loadFromGithubRepo}},
  \code{\link{loadFromLocalRepo}}; \code{\link{md5hash}};
  \code{\link{rmFromRepo}}; \code{\link{saveToRepo}};
  \code{\link{searchInGithubRepo}},
  \code{\link{searchInLocalRepo}};
  \code{\link{summaryGithubRepo}},
  \code{\link{summaryLocalRepo}}
}

