% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aread.R
\name{aread}
\alias{aread}
\title{Read Artifact Given as a md5hash from a Repository}
\usage{
aread(md5hash)
}
\arguments{
\item{md5hash}{One from following:

A character with at least three components sepearated with '/': GitHub user name, GitHub repository and name of the artifact (MD5 hash) or it's abbreviation.

A MD5 hash of object in current local default directory or it's abbreviation.}
}
\description{
\code{aread} reads an artifact from a Github \link{Repository}. It's a wrapper around 
\link{loadFromGithubRepo}.
}
\details{
Function \code{aread} read artifact (by the \code{md5hash}) from GitHub Repository.
It uses the function \link{loadFromGithubRepo} with different paramter's specification.
}
\examples{
\dontrun{
# read the object from local directory
setLocalRepo(system.file("graphGallery", package = "archivist"))
pl <- aread("2166dfbd3a7a68a91a2f8e6df1a44111")
# plot it
pl
# read the object from GitHub
pl <- aread("pbiecek/graphGallery/2166dfbd3a7a68a91a2f8e6df1a44111")
# plot it
pl
}
}
\author{
Przemyslaw Biecek, \email{przemyslaw.biecek@gmail.com}
}
\seealso{
Other archivist: \code{\link{Repository}},
  \code{\link{Tags}}, \code{\link{\%a\%}},
  \code{\link{addTagsRepo}}, \code{\link{ahistory}},
  \code{\link{aoptions}},
  \code{\link{archivist-github-integration}},
  \code{\link{archivist-package}}, \code{\link{asearch}},
  \code{\link{cache}}, \code{\link{copyLocalRepo}},
  \code{\link{createEmptyRepo}}, \code{\link{deleteRepo}},
  \code{\link{getTagsLocal}},
  \code{\link{loadFromLocalRepo}}, \code{\link{md5hash}},
  \code{\link{rmFromRepo}}, \code{\link{saveToRepo}},
  \code{\link{searchInLocalRepo}},
  \code{\link{setLocalRepo}},
  \code{\link{shinySearchInLocalRepo}},
  \code{\link{showLocalRepo}},
  \code{\link{summaryLocalRepo}},
  \code{\link{zipLocalRepo}}
}

