% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloneGithubRepo.R
\name{cloneGithubRepo}
\alias{cloneGithubRepo}
\title{Clone Github Repository}
\usage{
cloneGithubRepo(repoURL, repoDir = NULL, ...)
}
\arguments{
\item{repoURL}{The remote repository to clone.}

\item{repoDir}{Local directory to clone to. If \code{NULL}, by default, creates a local directory.
which corresponds to the name after last \code{/} in \code{repoURL}.}

\item{...}{Further parameters passed to \link[git2r]{clone}.}
}
\description{
\code{cloneGithubRepo} is a wrapper around \code{git clone} and clones GitHub Repository
into the \code{local_path} directory.

More archivist functionalities that integrate archivist and GitHub API can be found here \link{archivist-github-integration} (\link{agithub}).
}
\examples{
\dontrun{

cloneGithubRepo("https://github.com/MarcinKosinski/Museum")
cloneGithubRepo("https://github.com/MarcinKosinski/Museum-Extra")


# empty Github Repository creation

library(httr)
myapp <- oauth_app("github",
                   key = app_key,
                   secret = app_secret)
github_token <- oauth2.0_token(oauth_endpoints("github"),
                               myapp,
                               scope = "public_repo")
# setting options                              
aoptions("github_token", github_token)
aoptions("user.name", user_name)
aoptions("user.password", user_password)

createEmptyGithubRepo("archive-test4")
setGithubRepo(aoptions("user.name"), "archive-test4")
## artifact's archiving
przyklad <- 1:100

# archiving
archive(przyklad) -> md5hash_path

## proof that artifact is really archived
showGithubRepo() # uses options from setGithubRepo
# let's remove przyklad
rm(przyklad)
# and load it back from md5hash_path
aread(md5hash_path)


# clone example
unlink("archive-test", recursive = TRUE)
cloneGithubRepo('https://github.com/MarcinKosinski/archive-test')
setGithubRepo(aoptions("user.name"), "archive-test")
data(iris)
archive(iris)
showGithubRepo()


}
}
\author{
Marcin Kosinski, \email{m.p.kosinski@gmail.com}
}
\seealso{
Other archivist: \code{\link{Repository}},
  \code{\link{Tags}}, \code{\link{\%a\%}},
  \code{\link{addTagsRepo}}, \code{\link{ahistory}},
  \code{\link{aoptions}}, \code{\link{archive}},
  \code{\link{archivist-github-integration}},
  \code{\link{archivist-package}}, \code{\link{aread}},
  \code{\link{asearch}}, \code{\link{cache}},
  \code{\link{copyLocalRepo}},
  \code{\link{createEmptyRepo}}, \code{\link{deleteRepo}},
  \code{\link{getTagsLocal}},
  \code{\link{loadFromLocalRepo}}, \code{\link{md5hash}},
  \code{\link{rmFromRepo}}, \code{\link{saveToRepo}},
  \code{\link{searchInLocalRepo}},
  \code{\link{setLocalRepo}},
  \code{\link{shinySearchInLocalRepo}},
  \code{\link{showLocalRepo}},
  \code{\link{splitTagsLocal}},
  \code{\link{summaryLocalRepo}},
  \code{\link{zipLocalRepo}}
}

