% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloneGithubRepo.R
\name{cloneGithubRepo}
\alias{cloneGithubRepo}
\title{Clone Github Repository}
\usage{
cloneGithubRepo(repoURL, local_path = NULL, ...)
}
\arguments{
\item{repoURL}{The remote repository to clone.}

\item{local_path}{Local directory to clone to. If \code{NULL}, by default, creates a local directory
which corresponds to the name after last \code{/} in \code{repoURL}.}

\item{...}{Further parameters passed to \link[git2r]{clone}.}
}
\description{
\code{cloneGithubRepo} is a wrapper around \code{git clone} and clones GitHub Repository
into the \code{local_path} directory.

More archivist functionalities that integrate archivist and GitHub API can be found here \link{archivist-github-integration}.
}
\examples{
\dontrun{

cloneGithubRepo("https://github.com/MarcinKosinski/Museum")
cloneGithubRepo("https://github.com/MarcinKosinski/Museum-Extra")

}
}
\author{
Marcin Kosinski, \email{m.p.kosinski@gmail.com}
}
\seealso{
Other archivist: \code{\link{Repository}},
  \code{\link{Tags}}, \code{\link{\%a\%}},
  \code{\link{addTagsRepo}}, \code{\link{ahistory}},
  \code{\link{aoptions}}, \code{\link{archive}},
  \code{\link{archivist-github-integration}},
  \code{\link{archivist-package}}, \code{\link{aread}},
  \code{\link{asearch}}, \code{\link{cache}},
  \code{\link{copyLocalRepo}},
  \code{\link{createEmptyRepo}}, \code{\link{deleteRepo}},
  \code{\link{getTagsLocal}},
  \code{\link{loadFromLocalRepo}}, \code{\link{md5hash}},
  \code{\link{rmFromRepo}}, \code{\link{saveToRepo}},
  \code{\link{searchInLocalRepo}},
  \code{\link{setLocalRepo}},
  \code{\link{shinySearchInLocalRepo}},
  \code{\link{showLocalRepo}},
  \code{\link{summaryLocalRepo}},
  \code{\link{zipLocalRepo}}
}

