% Generated by roxygen2 (4.0.1): do not edit by hand
\name{createEmptyRepo}
\alias{createEmptyRepo}
\title{Create an Empty Repository in Given Directory}
\usage{
createEmptyRepo(dir)
}
\arguments{
\item{dir}{A character that specifies the directory for repository to be made.}
}
\description{
\code{createEmptyRepo} creates an empty \link{Repository} for given directory in which archivised objects will be stored.
}
\details{
At least one repository must be initialized prior to use other functions from archivist package.
If working in groups, it is highly recommend to create repository on shared Dropbox/Git folder.

All objects desired to be archivised are going to be saved in the local Repository, which is a SQLite database stored in a file named \code{backpack}.
Every object is saved (after calling \code{saveToRepo} function) in a \code{md5hash.rda} file, located in the folder (created in given directory) named
\code{gallery}. \code{md5hash} is a hash generated from object, which is wanted to be saved and is different
for various objects. This \code{md5hash} is a string of length 32 that comes out as a result of
\code{digest{digest}} function, which uses a cryptographical MD5 hash algorithm.

Created \code{backpack} database is a useful and fundamental tool for remembering object's
\code{name}, \code{class}, \code{archivisation date} etc (that are remembered as \link{Tags})
or for keeping object's \code{md5hash}.

Besides \code{backpack} database, there also is created a folder named \code{gallery} in which all
objects will be archivised.

After every \code{saveToRepo} call the database is refreshed, so object is available immediately.
}
\examples{
exampleDir <- tempdir()
createEmptyRepo(dir = exampleDir)

# check the state of empty Repository

summaryRepo( method = "objects", dir = exampleDir )
summaryRepo( method = "tags", dir = exampleDir )

# removing all files generated to this function's examples
file.remove( paste0( exampleDir, "/backpack.db" ) )
file.remove( paste0( exampleDir, "/gallery" ) )
}
\author{
Marcin Kosinski, \email{m.p.kosinski@gmail.com}
}
\seealso{
Other archivist: \code{\link{Repository}};
  \code{\link{Tags}}; \code{\link{archivist-package}};
  \code{\link{loadFromGithubRepo}},
  \code{\link{loadFromLocalRepo}}; \code{\link{md5hash}};
  \code{\link{rmFromRepo}}; \code{\link{saveToRepo}};
  \code{\link{searchInGithubRepo}},
  \code{\link{searchInLocalRepo}};
  \code{\link{summaryGithubRepo}},
  \code{\link{summaryLocalRepo}}
}

