% Generated by roxygen2 (4.0.1): do not edit by hand
\name{createEmptyRepo}
\alias{createEmptyRepo}
\title{Create an Empty Repository in a Given Directory}
\usage{
createEmptyRepo(repoDir, force = FALSE)
}
\arguments{
\item{repoDir}{A character that specifies the directory for the Repository to be made.}

\item{force}{If \code{force = TRUE} function call forces to create \code{repoDir} directory if
it did not exist. Default set to \code{force = FALSE}.}
}
\description{
\code{createEmptyRepo} creates an empty \link{Repository} in a given directory in which archived objects will be stored.
}
\details{
At least one Repository must be initialized before using other functions from the \pkg{archivist} package.
When working in groups, it is highly recommended to create a Repository on a shared Dropbox/Git folder.

All objects desired to be archived are going to be saved in the local Repository, which is an SQLite database stored in a file named \code{backpack}.
After calling \code{saveToRepo} function, every object will be archived in a \code{md5hash.rda} file. This file will be saved in a folder (under \code{repoDir} directory) named
\code{gallery}. For every object, \code{md5hash} is a unique string of length 32 that comes out as a result of
\code{digest{digest}} function, which uses a cryptographical MD5 hash algorithm.

Created \code{backpack} database is a useful and fundamental tool for remembering object's
\code{name}, \code{class}, \code{archiving date} etc. (that are remembered as \link{Tags}),
or for keeping object's \code{md5hash}.

Besides the \code{backpack} database, \code{gallery} folder is created in which all
objects will be archived.

After every \code{saveToRepo} call the database is refreshed, so an object is available
immediately in \code{backpack.db} database for other collaborators.
}
\examples{
\dontrun{
exampleRepoDir <- tempdir()
createEmptyRepo( repoDir = exampleRepoDir )

# check the state of an empty Repository

summaryLocalRepo( method = "md5hashes", repoDir = exampleRepoDir )
summaryLocalRepo( method = "tags", repoDir = exampleRepoDir )

# removing all files generated to this function's examples
file.remove( paste0( exampleRepoDir, "/backpack.db" ) )

rm( exampleRepoDir )
}
}
\author{
Marcin Kosinski, \email{m.p.kosinski@gmail.com}
}
\seealso{
Other archivist: \code{\link{Repository}};
  \code{\link{Tags}}; \code{\link{archivist-package}};
  \code{\link{copyGithubRepo}},
  \code{\link{copyLocalRepo}}; \code{\link{deleteRepo}};
  \code{\link{infoGithubRepo}},
  \code{\link{infoLocalRepo}};
  \code{\link{loadFromGithubRepo}},
  \code{\link{loadFromLocalRepo}}; \code{\link{md5hash}};
  \code{\link{rmFromRepo}}; \code{\link{saveToRepo}};
  \code{\link{searchInGithubRepo}},
  \code{\link{searchInLocalRepo}};
  \code{\link{showGithubRepo}}, \code{\link{showLocalRepo}}
}

