% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRemoteHook.R
\name{getRemoteHook}
\alias{getRemoteHook}
\title{Get http Hook for Remote Repo}
\usage{
getRemoteHook(repo = NULL, user = NULL, branch = "master",
  repoDirGit = FALSE, type = "github")
}
\arguments{
\item{repo}{A character containing a name of a Git repository on which the Repository is archived.}

\item{user}{A character containing a name of a Git user on whose account the \code{repo} is created.}

\item{branch}{A character containing a name of Git Repository's branch on which the Repository is archived. 
Default \code{branch} is \code{master}.}

\item{repoDirGit}{A character containing a name of a directory on Git repository 
on which the Repository is stored. If the Repository is stored in main folder on Git repository, this should be set 
to \code{repoDirGit = FALSE} as default.}

\item{type}{A character containing a type of the remote repository. Currently it can be 'github' or 'bitbucket'.}
}
\description{
\code{getRemoteHook} returns http adress of the remote \link{Repository}.
Then it can be used to download artifacts from the remote \link{Repository}.
}
\examples{

\dontrun{
# objects preparation
getRemoteHook("graphGallery", "pbiecek")
}
}
\author{
Przemyslaw Biecek, \email{przemyslaw.biecek@gmail.com}
}
\seealso{
Other archivist: \code{\link{Repository}},
  \code{\link{Tags}}, \code{\link{\%a\%}},
  \code{\link{addHooksToPrint}}, \code{\link{addTagsRepo}},
  \code{\link{ahistory}}, \code{\link{alink}},
  \code{\link{aoptions}}, \code{\link{archive}},
  \code{\link{archivist-github-integration}},
  \code{\link{archivist-package}}, \code{\link{aread}},
  \code{\link{asearch}}, \code{\link{cache}},
  \code{\link{cloneGithubRepo}},
  \code{\link{copyLocalRepo}},
  \code{\link{createEmptyRepo}},
  \code{\link{createGithubMDGallery}},
  \code{\link{deleteRepo}}, \code{\link{getTagsLocal}},
  \code{\link{loadFromLocalRepo}}, \code{\link{md5hash}},
  \code{\link{pushRepo}}, \code{\link{rmFromRepo}},
  \code{\link{saveToRepo}},
  \code{\link{searchInLocalRepo}},
  \code{\link{setLocalRepo}},
  \code{\link{shinySearchInLocalRepo}},
  \code{\link{showLocalRepo}},
  \code{\link{splitTagsLocal}},
  \code{\link{summaryLocalRepo}},
  \code{\link{zipLocalRepo}}
}

