% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getTags.R
\name{getTagsLocal}
\alias{getTagsGithub}
\alias{getTagsLocal}
\title{Return a Tag Corresponding to md5hash}
\usage{
getTagsLocal(md5hash, repoDir = NULL, tag = "name")

getTagsGithub(md5hash, user = NULL, repo = NULL, branch = "master",
  repoDirGit = FALSE, tag = "name")
}
\arguments{
\item{md5hash}{A character containing \code{md5hash}
of artifacts which \code{Tag} is desired to be returned.}

\item{repoDir}{A character denoting an existing directory in
which an artifact is stored. If set to \code{NULL} (by default), uses the \code{repoDir} specified in \link{setLocalRepo}.}

\item{tag}{A type of a \link{Tags}. Default \code{tag = "name"}.}

\item{user}{Only if working with a Github repository. A character containing a name of a Github user on whose account the \code{repo} is created.
By default set to \code{NULL} - see \code{Note}.}

\item{repo}{Only if working with a Github repository. A character containing a name of a Github repository on which the Repository is archived.
By default set to \code{NULL} - see \code{Note}.}

\item{branch}{Only if working with a Github repository. A character containing a name of
Github repository's branch in which Repository is archived. Default \code{branch} is \code{master}.}

\item{repoDirGit}{Only if working with a Github repository. A character containing a name of a directory on Github repository
on which the Repository is stored. If the Repository is stored in main folder on Github repository, this should be set
to \code{repoDirGit = FALSE} as default.}
}
\value{
The character is returned, which is a \code{Tag} (see \link{Tags}) related to \link{md5hash}
of an artifact.
}
\description{
\code{getTagsLocal} and \code{getTagsGithub} return a \code{Tag} (see \link{Tags}) related to \link{md5hash}
of an artifact. To learn more about artifacts visit \link[archivist]{archivist-package}.
}
\details{
\code{getTagsLocal} and \code{getTagsGithub} return (see \link{Tags}) related to \link{md5hash}
of an artifact. To learn more about artifacts visit \link[archivist]{archivist-package}.
}
\note{
If \code{repo} and \code{user} are set to \code{NULL} (as default) in Github mode then global parameters
set in \link{setGithubRepo} function are used.
}
\examples{
\dontrun{
library(dplyr)
exampleRepoDir <- tempdir()
createEmptyRepo( exampleRepoDir )

data(mtcars)
hash <- mtcars \%.\%
 group_by(cyl, am) \%.\%
 select(mpg, cyl, wt, am) \%.\%
 summarise(avgmpg = mean(mpg), avgwt = mean(wt)) \%.\%
 filter(avgmpg > 20) \%>\%
 saveToRepo( exampleRepoDir )

getTagsLocal( md5hash = hash, exampleRepoDir )

deleteRepo( exampleRepoDir, TRUE)
rm( exampleRepoDir )

getTagsGithub( "3db63bc63b8defaf42c0bde19160f242",
   user="pbiecek", repo="archivist")

# many archivist-like Repositories on one Github repository

getTagsGithub("ff575c261c949d073b2895b05d1097c3", user="MarcinKosinski",
repo="Museum", branch="master", repoDirGit="ex1")

}
}
\author{
Marcin Kosinski, \email{m.p.kosinski@gmail.com}
}
\seealso{
Other archivist: \code{\link{Repository}};
  \code{\link{Tags}}; \code{\link{\%a\%}};
  \code{\link{addTagsRepo}}; \code{\link{ahistory}},
  \code{\link{print.ahistory}},
  \code{\link{print.ahistoryKable}};
  \code{\link{aoptions}}; \code{\link{archivist-package}};
  \code{\link{aread}}; \code{\link{asearch}};
  \code{\link{cache}}; \code{\link{copyGithubRepo}},
  \code{\link{copyLocalRepo}};
  \code{\link{createEmptyGithubRepo}};
  \code{\link{createEmptyRepo}}; \code{\link{deleteRepo}};
  \code{\link{loadFromGithubRepo}},
  \code{\link{loadFromLocalRepo}}; \code{\link{md5hash}};
  \code{\link{multiSearchInGithubRepo}},
  \code{\link{multiSearchInLocalRepo}},
  \code{\link{searchInGithubRepo}},
  \code{\link{searchInLocalRepo}};
  \code{\link{rmFromRepo}}; \code{\link{saveToRepo}};
  \code{\link{setGithubRepo}}, \code{\link{setLocalRepo}};
  \code{\link{shinySearchInLocalRepo}};
  \code{\link{showGithubRepo}},
  \code{\link{showLocalRepo}};
  \code{\link{summaryGithubRepo}},
  \code{\link{summaryLocalRepo}};
  \code{\link{zipGithubRepo}}, \code{\link{zipLocalRepo}}

Other archivist: \code{\link{Repository}};
  \code{\link{Tags}}; \code{\link{\%a\%}};
  \code{\link{addTagsRepo}}; \code{\link{ahistory}},
  \code{\link{print.ahistory}},
  \code{\link{print.ahistoryKable}};
  \code{\link{aoptions}}; \code{\link{archivist-package}};
  \code{\link{aread}}; \code{\link{asearch}};
  \code{\link{cache}}; \code{\link{copyGithubRepo}},
  \code{\link{copyLocalRepo}};
  \code{\link{createEmptyGithubRepo}};
  \code{\link{createEmptyRepo}}; \code{\link{deleteRepo}};
  \code{\link{loadFromGithubRepo}},
  \code{\link{loadFromLocalRepo}}; \code{\link{md5hash}};
  \code{\link{multiSearchInGithubRepo}},
  \code{\link{multiSearchInLocalRepo}},
  \code{\link{searchInGithubRepo}},
  \code{\link{searchInLocalRepo}};
  \code{\link{rmFromRepo}}; \code{\link{saveToRepo}};
  \code{\link{setGithubRepo}}, \code{\link{setLocalRepo}};
  \code{\link{shinySearchInLocalRepo}};
  \code{\link{showGithubRepo}},
  \code{\link{showLocalRepo}};
  \code{\link{summaryGithubRepo}},
  \code{\link{summaryLocalRepo}};
  \code{\link{zipGithubRepo}}, \code{\link{zipLocalRepo}}
}

