% Generated by roxygen2 (4.0.1): do not edit by hand
\name{loadFromLocalRepo}
\alias{loadFromGithubRepo}
\alias{loadFromLocalRepo}
\title{Load Object Given as md5hash from Repository}
\usage{
loadFromLocalRepo(md5hash, dir)

loadFromGithubRepo(md5hash, repo, user)
}
\arguments{
\item{dir}{A character denoting an existing directory from which an object will be loaded.}

\item{md5hash}{A hash of an object. A character string being a result of a cryptographical hash function with MD5 algorithm.}

\item{repo}{Only if working on Github Repository. A character string containing a name of Github Repository.}

\item{user}{Only if working on Github Repository. A character string containing a name of Github User.}
}
\description{
\code{loadFromLocalRepo} loads an object from a local repository into workspace.
\code{loadFromGithubRepo} loads an object from a github repository into workspace.
}
\details{
Those functions load objects from local or Github Repositories. Both of them take \code{md5hash} as a
parameter. This \code{md5hash} is a string of length 32 that comes out as a result from \link{saveToRepo} function, which uses a cryptographical hash function with MD5 algorithm

Note that \code{user} and \code{repo} should be used only when working on Github Repository and ought to be ommited in a local working mode,
when \code{dir} should only be used when working on a local Repository and ought to be ommited in a Github working mode.
}
\examples{
# not work
library(digest)
loadFromLocalRepo( md5hash = digest(plot123) , dir = getwd() )
loadFromLocalRepo( md5hash = "hs82h9kahs72h3nckv0dhsu28sgd73js",
                       dir = "user/folder/here" )
loadFromGithubRepo( md5hash = "hs82h9kahs72h3nckv0dhsu28sgd73js",
                           repo = "REPONAME", user = "USERNAME" )
}
\author{
autor
}
\seealso{
Other archivist: \code{\link{createEmptyRepo}};
  \code{\link{rmFromRepo}}; \code{\link{saveToRepo}};
  \code{\link{searchInGithubRepo}},
  \code{\link{searchInLocalRepo}}
}

