% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{md5hash}
\alias{md5hash}
\title{md5hash}
\description{
\code{Repository} stores specific values of an object, different for
various object's classes, and objects themselves. Objects are archived with
a special attribute named \code{md5hash}.
}
\details{
For every object, \code{md5hash} is a unique string of length 32 that comes out as a result of
\code{digest{digest}} function, which uses a cryptographical MD5 hash algorithm.
The \code{md5hash} of every object, that is archived to the \link{Repository}, is also saved
to the Repository with this object's \code{Tags} - see \link{Tags}. It enables to distinguish
objects in Repository and facilitates searching and loading them.
}
\seealso{
Functions that take \code{md5hash} as a parameter are:
\itemize{
 \item \link{rmFromRepo},
 \item \link{loadFromLocalRepo},
 \item \link{loadFromGithubRepo},
}
Functions returning \code{md5hash} as a value are:
\itemize{
 \item \link{searchInLocalRepo},
 \item \link{searchInGithubRepo},
 \item \link{saveToRepo}.
}
Functions returning \code{md5hahes} as a data.frame are:
\itemize{
 \item \link{summaryLocalRepo},
 \item \link{summaryGithubRepo}.
}

Other archivist: \code{\link{Repository}};
  \code{\link{Tags}}; \code{\link{archivist-package}};
  \code{\link{copyGithubRepo}},
  \code{\link{copyLocalRepo}};
  \code{\link{createEmptyRepo}}; \code{\link{deleteRepo}};
  \code{\link{infoGithubRepo}},
  \code{\link{infoLocalRepo}};
  \code{\link{loadFromGithubRepo}},
  \code{\link{loadFromLocalRepo}};
  \code{\link{rmFromRepo}}; \code{\link{saveToRepo}};
  \code{\link{searchInGithubRepo}},
  \code{\link{searchInLocalRepo}};
  \code{\link{summaryGithubRepo}},
  \code{\link{summaryLocalRepo}}
}

