% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push_pull_Repo.R
\name{pushRepo}
\alias{pullRepo}
\alias{pushRepo}
\title{Push and Pull for Local Repository Synchronized with GitHub}
\usage{
pushRepo(repoDir = aoptions("repoDir"),
  commitMessage = aoptions("commitMessage"), repo = aoptions("repo"),
  user = aoptions("user"), password = aoptions("password"),
  files = c("gallery", "backpack.db"))

pullRepo(repoDir = aoptions("repoDir"), user = aoptions("user"),
  password = aoptions("password"), ...)
}
\arguments{
\item{repoDir}{A character specifing the directory to Local \code{Repository} to/from which artifacts will be pulled/pushed to GitHub.}

\item{commitMessage}{A character denoting a message added to the commit while performing push.
By default specified to \code{NULL} which corresponds to commit message \code{archivist: pushRepo}.}

\item{repo}{A character denoting GitHub repository name and synchronized local existing directory in which an artifact will be saved.}

\item{user}{A character denoting GitHub user name. Can be set globally with \code{aoptions("user", user)}.
See \link{archivist-github-integration}.}

\item{password}{A character denoting GitHub user password. Can be set globally with \code{aoptions("password", password)}.
See \link{archivist-github-integration}.}

\item{files}{A character vector containing directories to files that should be commited and pushed. The working directory
is \code{repoDir}. By default all uncommited artifacts and \code{backpack.db} will be pushed.}

\item{...}{Further arguments passed to \link[git2r]{push} or \link[git2r]{pull}.}
}
\description{
\code{pushRepo} adds files, commits them and pushes from Local \link{Repository} to synchronized GitHub one. 
\code{pullRepo} pulls (\code{git pull}) changes from remote GitHub \code{Repository} to the correspoding Local one.
More archivist functionalities that integrate archivist and GitHub API can be found here \link{archivist-github-integration} (\link{agithub}).
}
\note{
To check the \code{status} (\code{git status}) of the Repository use \code{git2r::status(repository(repoDir))}. See examples.
}
\examples{
\dontrun{

library(httr)
myapp <- oauth_app("github",
                   key = app_key,
                   secret = app_secret)
github_token <- oauth2.0_token(oauth_endpoints("github"),
                                myapp,
                                scope = c("public_repo",
                                          "delete_repo"))
aoptions("github_token", github_token)
aoptions("name", user.name)
aoptions("password", user.password)

createEmptyGithubRepo("Museum", default = TRUE) # here github_token is used
data(iris)
saveToRepo(iris)
git2r::status(repository('Museum'))
pushRepo(commitMessage = "add iris")
git2r::status(repository('Museum'))

}
}
\author{
Marcin Kosinski, \email{m.p.kosinski@gmail.com}
}
\seealso{
Other archivist: \code{\link{Repository}},
  \code{\link{Tags}}, \code{\link{\%a\%}},
  \code{\link{addTagsRepo}}, \code{\link{ahistory}},
  \code{\link{alink}}, \code{\link{aoptions}},
  \code{\link{archive}},
  \code{\link{archivist-github-integration}},
  \code{\link{archivist-package}}, \code{\link{aread}},
  \code{\link{asearch}}, \code{\link{cache}},
  \code{\link{cloneGithubRepo}},
  \code{\link{copyLocalRepo}},
  \code{\link{createEmptyRepo}},
  \code{\link{createGithubMDGallery}},
  \code{\link{deleteRepo}},
  \code{\link{getRemoteHookGithub}},
  \code{\link{getTagsLocal}},
  \code{\link{loadFromLocalRepo}}, \code{\link{md5hash}},
  \code{\link{rmFromRepo}}, \code{\link{saveToRepo}},
  \code{\link{searchInLocalRepo}},
  \code{\link{setLocalRepo}},
  \code{\link{shinySearchInLocalRepo}},
  \code{\link{showLocalRepo}},
  \code{\link{splitTagsLocal}},
  \code{\link{summaryLocalRepo}},
  \code{\link{zipLocalRepo}}
}

