% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{Repository}
\alias{Repository}
\title{Repository}
\description{
\code{Repository} stores specific values of an object, different for
various object's classes, and objects themselves.
}
\details{
\code{Repository} is folder with a SQLite database stored in a file named \code{backpack}
and a subdirectory named \code{gallery} with collection of object saved as \code{.rda} files.
}
\seealso{
Functions using \code{Repository} are:
\itemize{
 \item \link{saveToRepo},
 \item \link{rmFromRepo},
 \item \link{loadFromLocalRepo},
 \item \link{loadFromGithubRepo},
 \item \link{searchInLocalRepo},
 \item \link{searchInGithubRepo},
 \item \link{summaryLocalRepo},
 \item \link{summaryGithubRepo}.
 }

Function creating \code{Repository} is:
\link{createEmptyRepo}.

Other archivist: \code{\link{Tags}};
  \code{\link{archivist-package}};
  \code{\link{createEmptyRepo}};
  \code{\link{loadFromGithubRepo}},
  \code{\link{loadFromLocalRepo}}; \code{\link{md5hash}};
  \code{\link{rmFromRepo}}; \code{\link{saveToRepo}};
  \code{\link{searchInGithubRepo}},
  \code{\link{searchInLocalRepo}};
  \code{\link{summaryGithubRepo}},
  \code{\link{summaryLocalRepo}}
}

