% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restoreLibraries.R
\name{restoreLibs}
\alias{restoreLibs}
\title{Restore Versions of Libraries}
\usage{
restoreLibs(md5hash, session_info = NULL, lib.loc = NULL)
}
\arguments{
\item{md5hash}{One of the following (see \link{aread}):

A character vector which elements are consisting of at least three components separated 
with '/': Remote user name, Remote repository and name of the artifact (MD5 hash) or 
it's abbreviation.

MD5 hashes of artifacts in current local default directory or its abbreviations.}

\item{session_info}{Object with versions of packages to be installed. If not supplied then 
it will be extracted from md5hash \code{ md5hash }}

\item{lib.loc}{A character vector describing the location of \code{R} library trees to restore archived libraries, or \code{NULL}.
The default value of \code{NULL} corresponds to all libraries currently known to \link{.libPaths}(). 
Non-existent library trees are silently ignored.}
}
\description{
Function \code{ restoreLibs } gets either \code{session info} or artifact's \code{md5hash} 
and restore libraries/packages to versions attached when the object was saved in the repo. 
Typical use case is following. We have saved an object and now we are restoring it, but
with current version of packages something is not working. The function \code{restoreLibs()} 
reverts all libraries that were attached previously to their previous versions.
}
\section{Contact}{

 
Bug reports and feature requests can be sent to 
\href{https://github.com/pbiecek/archivist/issues}{https://github.com/pbiecek/archivist/issues}
}

\examples{
\dontrun{
## objects preparation
restoreLibs(md5hash = "pbiecek/graphGallery/7f3453331910e3f321ef97d87adb5bad")
}
}
\references{
Przemyslaw Biecek, Marcin Kosinski (2017). 
archivist: An R Package for Managing, Recording and Restoring Data Analysis Results 
Conditionally accepted for Journal of Statistical Software
URL https://github.com/pbiecek/archivist
}
\seealso{
Other archivist: \code{\link{Repository}},
  \code{\link{Tags}}, \code{\link{\%a\%}},
  \code{\link{addHooksToPrint}}, \code{\link{addTagsRepo}},
  \code{\link{aformat}}, \code{\link{ahistory}},
  \code{\link{alink}}, \code{\link{aoptions}},
  \code{\link{archivist-package}}, \code{\link{aread}},
  \code{\link{asearch}}, \code{\link{asession}},
  \code{\link{atrace}}, \code{\link{cache}},
  \code{\link{copyLocalRepo}},
  \code{\link{createLocalRepo}},
  \code{\link{createMDGallery}},
  \code{\link{deleteLocalRepo}},
  \code{\link{getRemoteHook}}, \code{\link{getTagsLocal}},
  \code{\link{loadFromLocalRepo}}, \code{\link{md5hash}},
  \code{\link{removeTagsRepo}},
  \code{\link{rmFromLocalRepo}},
  \code{\link{saveToLocalRepo}},
  \code{\link{searchInLocalRepo}},
  \code{\link{setLocalRepo}},
  \code{\link{shinySearchInLocalRepo}},
  \code{\link{showLocalRepo}},
  \code{\link{splitTagsLocal}},
  \code{\link{summaryLocalRepo}},
  \code{\link{zipLocalRepo}}
}
\author{
Marcin Kosinski, \email{m.p.kosinski@gmail.com} \\
Przemyslaw Biecek, \email{przemyslaw.biecek@gmail.com}
}
