% Generated by roxygen2 (4.0.1): do not edit by hand
\name{returnTag}
\alias{returnTag}
\title{Returen a \code{Tag} Corresponding to \code{md5hash}}
\usage{
returnTag(md5hash, repoDir, tag = "name")
}
\arguments{
\item{repoDir}{A character denoting an existing directory in
which an artifact is stored.}

\item{md5hashes}{A character containing \code{md5hash}
of artifacts which \code{Tag} is desired to be returned.}

\item{tag}{A type of a \link{Tags}. Default \code{tag = "name"}.}
}
\value{
A tag.
}
\description{
\code{returnTag}
}
\details{
\code{returnTag}
}
\examples{
\dontrun{
}
}
\author{
Marcin Kosinski, \email{m.p.kosinski@gmail.com}
}
\seealso{
Other archivist: \code{\link{Repository}};
  \code{\link{Tags}}; \code{\link{archivist-package}};
  \code{\link{copyGithubRepo}},
  \code{\link{copyLocalRepo}};
  \code{\link{createEmptyRepo}}; \code{\link{deleteRepo}};
  \code{\link{loadFromGithubRepo}},
  \code{\link{loadFromLocalRepo}}; \code{\link{md5hash}};
  \code{\link{rmFromRepo}}; \code{\link{saveToRepo}};
  \code{\link{searchInGithubRepo}},
  \code{\link{searchInLocalRepo}};
  \code{\link{showGithubRepo}},
  \code{\link{showLocalRepo}};
  \code{\link{summaryGithubRepo}},
  \code{\link{summaryLocalRepo}}
}

