% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rmFromRepo}
\alias{rmFromRepo}
\title{Remove an Object from a Repository}
\usage{
rmFromRepo(md5hash, dir)
}
\arguments{
\item{md5hash}{A hash of an object. A character string being a result of a cryptographical hash function with MD5 algorithm.}

\item{dir}{A character denoting an existing directory from which an object will be removed.}
}
\description{
\code{rmFromRepo} removes an object from a Repository given as \code{md5hash} - a string of
length 32 that comes out as a result from \link{saveToRepo} function, which uses a cryptographical hash function with MD5 algorithm
}
\details{
\code{rmFromRepo} removes an object from a Repository given as \code{md5hash} - a string of
length 32 that comes out as a result from \link{saveToRepo} function, which uses a cryptographical hash function with MD5 algorithm
}
\examples{
#example
}
\author{
autor
}
\seealso{
Other archivist: \code{\link{createEmptyRepo}};
  \code{\link{loadFromGithubRepo}},
  \code{\link{loadFromLocalRepo}};
  \code{\link{saveToRepo}};
  \code{\link{searchInGithubRepo}},
  \code{\link{searchInLocalRepo}}
}

