% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rmFromRepo}
\alias{rmFromRepo}
\title{Remove an Object Given as \code{md5hash} from a Repository}
\usage{
rmFromRepo(md5hash, repoDir, removeData = FALSE, removeMiniature = FALSE,
  force = FALSE)
}
\arguments{
\item{md5hash}{A character assigned to the object as a result of a cryptographical hash function with MD5 algorithm, or it's abbreviation. This object will be removed.}

\item{repoDir}{A character denoting an existing directory from which an object will be removed.}

\item{removeData}{A logical value denoting whether to remove a data with the \code{object} specified by the \code{md5hash}.
Defualt \code{FALSE}.}

\item{removeMiniature}{A logical value denoting whether to remove a miniature with the \code{object} specified by the \code{md5hash}.
Defualt \code{FALSE}.}

\item{force}{A logical value denoting whether to remove data if it is related to more than 1 object.
Defualt \code{FALSE}.}
}
\description{
\code{rmFromRepo} removes an object given as \code{md5hash} from a \link{Repository}.
}
\details{
\code{rmFromRepo} removes an object given as \code{md5hash} from a Repository,
which is a SQLite database named \code{backpack} - created by a \link{createEmptyRepo} call.
For every object, \code{md5hash} is a unique string of length 32 that comes out as a result of
\code{digest{digest}} function, which uses a cryptographical MD5 hash algorithm.


Also this function removes a \code{md5hash.rda} file, where \code{md5hash} is the object's hash as above.


Important: instead of giving the whole \code{md5hash} character, the user can simply give first few characters of the \code{md5hash}.
For example, \code{a09dd} instead of \code{a09ddjdkf9kj33dcjdnfjgos9jd9jkcv}. All objects with the same corresponing \code{md5hash} abbreviation
will be removed from the \link{Repository} and from the \code{gallery} folder.

\code{rmFromRepo} provides functionality that enables to delete miniatures of the objects (.txt or .png files)
while removing .rda files. To disable this functionality use \code{removeMiniature = FALSE}. Also, if the
 data from the object were archived, the data will be removed by default but there is a possibility not to
 delete this data while performing \code{rmFromRepo} - simply use \code{removeData = TRUE}.


\code{rmFromRepo} provides functionality that enables to delete miniatures of the objects (.txt or .png files) while removing .rda files.
To delete miniature use \code{removeMiniature = TRUE}. Also if the data from the object was archived, there is a possibility to delete this
data while removing object that uses this data. Simply use \code{removeData = TRUE}.

If one wants to remove all objects created between two dates, it is suggested to
perform:
\itemize{
   \item \code{obj2rm <- searchInLocalRepo( tag = list(dateFrom, dateTo), repoDir = )}
   \item \code{sapply(obj2rm, rmFromRepo, repoDir = )}
}
}
\note{
\code{md5hash} can be a result of the \link{searchInRepo} function proceeded with \code{tag = NAME} argument,
where \code{NAME} is a tag that describes the property of the objects to be deleted.


For more information about \code{Tags} check \link{Tags}.
}
\examples{
# objects preparation
\dontrun{
# data.frame object
data(iris)

# ggplot/gg object
library(ggplot2)
df <- data.frame(gp = factor(rep(letters[1:3], each = 10)),y = rnorm(30))
library(plyr)
ds <- ddply(df, .(gp), summarise, mean = mean(y), sd = sd(y))
myplot123 <- ggplot(df, aes(x = gp, y = y)) +
  geom_point() +  geom_point(data = ds, aes(y = mean),
                             colour = 'red', size = 3)

# lm object
model <- lm(Sepal.Length~ Sepal.Width + Petal.Length + Petal.Width, data= iris)
model2 <- lm(Sepal.Length~ Sepal.Width + Petal.Width, data= iris)
model3 <- lm(Sepal.Length~ Sepal.Width, data= iris)

# agnes (twins) object
library(cluster)
data(votes.repub)
agn1 <- agnes(votes.repub, metric = "manhattan", stand = TRUE)

# fanny (partition) object
x <- rbind(cbind(rnorm(10, 0, 0.5), rnorm(10, 0, 0.5)),
           cbind(rnorm(15, 5, 0.5), rnorm(15, 5, 0.5)),
           cbind(rnorm( 3,3.2,0.5), rnorm( 3,3.2,0.5)))
fannyx <- fanny(x, 2)

# creating example Repository - that examples will work

exampleRepoDir <- tempdir()
createEmptyRepo(repoDir = exampleRepoDir)
myplo123Md5hash <- saveToRepo(myplot123, repoDir=exampleRepoDir)
irisMd5hash <- saveToRepo(iris, repoDir=exampleRepoDir)
modelMd5hash  <- saveToRepo(model, repoDir=exampleRepoDir)
agn1Md5hash <- saveToRepo(agn1, repoDir=exampleRepoDir)
fannyxMd5hash <- saveToRepo(fannyx, repoDir=exampleRepoDir)

# let's see how the Repository look like: summary
summaryLocalRepo(method = "md5hashes", repoDir = exampleRepoDir)
summaryLocalRepo(method = "tags", repoDir = exampleRepoDir)


# remove examples

rmFromRepo(fannyxMd5hash, repoDir = exampleRepoDir, removeData= FALSE)
# removeData = FALSE provides from removing archived "fannyxMd5hash object"-data from
# a Repository and gallery
rmFromRepo(irisMd5hash, repoDir = exampleRepoDir)
# not that also files in gallery folder, created in exampleRepoDir
# directory are being removed

# let's see how the Repository look like: summary
summaryLocalRepo(method = "md5hashes", repoDir = exampleRepoDir)
summaryLocalRepo(method = "tags", repoDir = exampleRepoDir)

# one can have the same object archivised in 3 different names
agn1Md5hash2 <- saveToRepo(agn1, repoDir=exampleRepoDir)
agn1Md5hash3 <- saveToRepo(agn1, repoDir=exampleRepoDir)

# md5hashes are the same for that same object (agn1)
agn1Md5hash == agn1Md5hash2
agn1Md5hash2 == agn1Md5hash3

# but there are 3 times more rows in Repository database (backpack.db).

# let's see how the Repository look like: summary

summaryLocalRepo(method = "md5hashes", repoDir = exampleRepoDir)
summaryLocalRepo(method = "tags", repoDir = exampleRepoDir)

# one easy call removes them all

rmFromRepo(agn1Md5hash, repoDir = exampleRepoDir, removeData = TRUE, removeMiniature = TRUE)
# removeMiniature = TRUE removes miniatures from gallery folder

# rest of object can be removed e.g. like this
# looking for dates of creation and then removing all objects
# from specific date

obj2rm <- searchInLocalRepo( tag = list(dateFrom = Sys.Date(), dateTo = Sys.Date()),
                             repoDir = exampleRepoDir )
sapply(obj2rm, rmFromRepo, repoDir = exampleRepoDir)
# above example removed all objects from Today

# let's see how the Repository look like: summary
summaryLocalRepo(method = "md5hashes", repoDir = exampleRepoDir)
summaryLocalRepo(method = "tags", repoDir = exampleRepoDir)

# one can also remove objects from only specific class
modelMd5hash  <- saveToRepo(model, repoDir=exampleRepoDir)
model2Md5hash  <- saveToRepo(model2, repoDir=exampleRepoDir)
model3Md5hash  <- saveToRepo(model3, repoDir=exampleRepoDir)
summaryLocalRepo(method = "md5hashes", repoDir = exampleRepoDir)
objMd5hash <- searchInLocalRepo("class:lm", repoDir = exampleRepoDir)
sapply(objMd5hash, rmFromRepo, repoDir = exampleRepoDir)
summaryLocalRepo(method = "md5hashes", repoDir = exampleRepoDir)

# note that this time data and miniatures were not removed


# once can remove object specifying only its md5hash abbreviation
myplo123Md5hash <- saveToRepo(myplot123, repoDir=exampleRepoDir)
summaryLocalRepo(method = "md5hashes", repoDir = exampleRepoDir)
# "84445ca3cb23ccf2656ff25cc0fcaccd"
# so example abbreviation might be : "84445ca3"
rmFromRepo("84445ca3", repoDir = exampleRepoDir)
summaryLocalRepo(method = "md5hashes", repoDir = exampleRepoDir)


# removing all files generated to this function's examples
x <- list.files( paste0( exampleRepoDir, "/gallery/" ) )
sapply( x , function(x ){
  file.remove( paste0( exampleRepoDir, "/gallery/", x ) )
})
file.remove( paste0( exampleRepoDir, "/backpack.db" ) )

rm( exampleRepoDir )
}
}
\author{
Marcin Kosinski , \email{m.p.kosinski@gmail.com}
}
\seealso{
Other archivist: \code{\link{Repository}};
  \code{\link{Tags}}; \code{\link{archivist-package}};
  \code{\link{copyGithubRepo}},
  \code{\link{copyLocalRepo}};
  \code{\link{createEmptyRepo}}; \code{\link{deleteRepo}};
  \code{\link{infoGithubRepo}},
  \code{\link{infoLocalRepo}};
  \code{\link{loadFromGithubRepo}},
  \code{\link{loadFromLocalRepo}}; \code{\link{md5hash}};
  \code{\link{saveToRepo}};
  \code{\link{searchInGithubRepo}},
  \code{\link{searchInLocalRepo}};
  \code{\link{summaryGithubRepo}},
  \code{\link{summaryLocalRepo}}
}

