% Generated by roxygen2 (4.0.1): do not edit by hand
\name{saveToRepo}
\alias{saveToRepo}
\title{Save an Object into a Repository}
\usage{
saveToRepo(object, repoDir, archiveData = TRUE, archiveTags = TRUE,
  archiveMiniature = TRUE, force = TRUE, rememberName = TRUE, ...)
}
\arguments{
\item{object}{An arbitrary R object to be saved. For supported objects see details.}

\item{...}{Graphical parameters denoting width and height of a miniature. See details.}

\item{archiveData}{A logical value denoting whether to archive the data from the \code{object}.}

\item{archiveTags}{A logical value denoting whether to archive tags from the \code{object}.}

\item{archiveMiniature}{A logical value denoting whether to archive a miniature of the \code{object}.}

\item{repoDir}{A character denoting an existing directory in which an object will be saved.}

\item{force}{A logical value denoting whether to archive \code{object} if it was already archived in
a Repository.}

\item{rememberName}{A logical value. Should not be changed by user. It is a technical parameter.}
}
\value{
As a result of this function a character string is returned, which determines
the \code{md5hash} of the object that was used in the \code{saveToRepo} function. If
\code{archiveData} was \code{TRUE}, the result also
has an attribute, named \code{data}, that determines \code{md5hash} of the data needed
to compute the object.
}
\description{
\code{saveToRepo} function saves desired objects to the local \link{Repository} in a given directory.
}
\details{
\code{saveToRepo} function saves desired objects to the local Repository in a given directory.
Objects are saved in the local Repository, which is a SQLite database named \code{backpack}.
After every \code{saveToRepo} call the database is refreshed, so the object is available immediately in the database.
Every object is archived in a \code{md5hash.rda} file. This file will be saved in a folder (under \code{repoDir} directory) named
\code{gallery}. For every object, \code{md5hash} is a unique string of length 32 that comes out as a result of
\code{digest{digest}} function, which uses a cryptographical MD5 hash algorithm.

By default, a miniature of an object and (if possible) a data set needed to compute this object are extracted.
They are also going to be saved in a file named by their \code{md5hash} in the \code{gallery} folder that exists in the directory
specified in the \code{repoDir} argument. Moreover, a specific \code{Tag}-relation is going to be added to the \code{backpack} dataset in case there is a need to load
the object with it's related data set - see \link{loadFromLocalRepo} or \link{loadFromGithubRepo}. Default settings
may be changed by using the \code{archiveData}, \code{archiveTag} or \code{archiveMiniature} arguments with the
\code{FALSE} value.

\code{Tags} are object's attributes, different for various object's classes. For more detailed
information check \link{Tags}

Archived object can be searched in the \code{backpack} dataset by using the
\link{searchInLocalRepo} or \link{searchInGithubRepo} functions. Objects can be searched by their \link{Tags},
\code{names}, \code{classes} or \code{archiving date}.

Graphical parameters.

If the object is of class \code{data.frame} or \code{archiveData = TRUE}, it is possible to specify
how many rows of that data should be archived by adding the argument \code{firstRows} with the n
specified number of rows. Note that, the date can be extracted only from the objects that are supported by the package; see \href{https://github.com/pbiecek/archivist/issues/5}{https://github.com/pbiecek/archivist/issues/5}

If the object is of class \code{lattice} or \code{ggplot}, and
\code{archiveMiniature = TRUE}, then it is
possible to set the miniature's \code{width} and \code{height} parameters. By default they are set to
\code{width = 800}, \code{height = 600}.

Supported object's classes are (so far):
\itemize{
 \item \code{lm},
 \item \code{data.frame},
 \item \code{ggplot},
 \item \code{htest},
 \item \code{trellis},
 \item \code{twins (result of agnes, diana or mona function)},
 \item \code{partition (result of pam, clara or fanny fuction)},
 \item \code{lda},
 \item \code{qda},
 \item \code{glmnet},
 \item \code{survfit}.
 }

To check what \code{Tags} will be extracted for various objects see \link{Tags}.
}
\examples{
# objects preparation
\dontrun{
# data.frame object
data(iris)

# ggplot/gg object
library(ggplot2)
df <- data.frame(gp = factor(rep(letters[1:3], each = 10)),y = rnorm(30))
library(plyr)
ds <- ddply(df, .(gp), summarise, mean = mean(y), sd = sd(y))
myplot123 <- ggplot(df, aes(x = gp, y = y)) +
  geom_point() +  geom_point(data = ds, aes(y = mean),
               colour = 'red', size = 3)

# lm object
model <- lm(Sepal.Length~ Sepal.Width + Petal.Length + Petal.Width, data= iris)

# agnes (twins) object
library(cluster)
data(votes.repub)
agn1 <- agnes(votes.repub, metric = "manhattan", stand = TRUE)

# fanny (partition) object
x <- rbind(cbind(rnorm(10, 0, 0.5), rnorm(10, 0, 0.5)),
         cbind(rnorm(15, 5, 0.5), rnorm(15, 5, 0.5)),
          cbind(rnorm( 3,3.2,0.5), rnorm( 3,3.2,0.5)))
fannyx <- fanny(x, 2)

# lda object
library(MASS)

Iris <- data.frame(rbind(iris3[,,1], iris3[,,2], iris3[,,3]),
                  Sp = rep(c("s","c","v"), rep(50,3)))
train <- c(8,83,115,118,146,82,76,9,70,139,85,59,78,143,68,
           134,148,12,141,101,144,114,41,95,61,128,2,42,37,
           29,77,20,44,98,74,32,27,11,49,52,111,55,48,33,38,
           113,126,24,104,3,66,81,31,39,26,123,18,108,73,50,
           56,54,65,135,84,112,131,60,102,14,120,117,53,138,5)
lda1 <- lda(Sp ~ ., Iris, prior = c(1,1,1)/3, subset = train)

# qda object
tr <- c(7,38,47,43,20,37,44,22,46,49,50,19,4,32,12,29,27,34,2,1,17,13,3,35,36)
train <- rbind(iris3[tr,,1], iris3[tr,,2], iris3[tr,,3])
cl <- factor(c(rep("s",25), rep("c",25), rep("v",25)))
qda1 <- qda(train, cl)

# glmnet object
library( glmnet )

zk=matrix(rnorm(100*20),100,20)
bk=rnorm(100)
glmnet1=glmnet(zk,bk)


# creating example Repository - that examples will work

# save examples

exampleRepoDir <- tempdir()
createEmptyRepo( repoDir = exampleRepoDir )
saveToRepo( myplot123, repoDir=exampleRepoDir )
saveToRepo( iris, repoDir=exampleRepoDir )
saveToRepo( model, repoDir=exampleRepoDir )
saveToRepo( agn1, repoDir=exampleRepoDir )
saveToRepo( fannyx, repoDir=exampleRepoDir )
saveToRepo( lda1, repoDir=exampleRepoDir )
saveToRepo( qda1, repoDir=exampleRepoDir )
saveToRepo( glmnet1, repoDir=exampleRepoDir )

# let's see how the Repository look like: summary

summaryLocalRepo( method = "md5hashes", repoDir = exampleRepoDir )
summaryLocalRepo( method = "tags", repoDir = exampleRepoDir )

# removing all files generated to this function's examples
x <- list.files( paste0( exampleRepoDir, "/gallery/" ) )
sapply( x , function(x ){
     file.remove( paste0( exampleRepoDir, "/gallery/", x ) )
   })
file.remove( paste0( exampleRepoDir, "/backpack.db" ) )

rm( exampleRepoDir )
}
}
\author{
Marcin Kosinski , \email{m.p.kosinski@gmail.com}
}
\seealso{
For more detailed information check package vignette - url needed.

The list of supported objects and their tags is available on \code{wikis} on \pkg{archivist} Github Repository:
\href{https://github.com/pbiecek/archivist/wiki/archivist-package---Tags}{https://github.com/pbiecek/archivist/wiki/archivist-package---Tags}.

Other archivist: \code{\link{Repository}};
  \code{\link{Tags}}; \code{\link{archivist-package}};
  \code{\link{createEmptyRepo}};
  \code{\link{infoGithubRepo}},
  \code{\link{infoLocalRepo}};
  \code{\link{loadFromGithubRepo}},
  \code{\link{loadFromLocalRepo}}; \code{\link{md5hash}};
  \code{\link{rmFromRepo}};
  \code{\link{searchInGithubRepo}},
  \code{\link{searchInLocalRepo}};
  \code{\link{summaryGithubRepo}},
  \code{\link{summaryLocalRepo}}
}

