% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setRepo.R
\name{setLocalRepo}
\alias{setLocalRepo}
\alias{setPostgresRepo}
\alias{setRemoteRepo}
\title{Set Repository's Global Path}
\usage{
setLocalRepo(repoDir)

setPostgresRepo(repoDir, connector = NULL)

setRemoteRepo(user, repo, branch = "master", subdir = "/",
  repoType = "github")
}
\arguments{
\item{repoDir}{A character denoting a directory of a Repository that we want to
make dafault.}

\item{connector}{If user want to use some external database instead of SQLite, then the \code{connector} shall be the function that create a \code{DBI} connection with the database.
Within every transaction the connection is opened and closed, thus the \code{connector} function will be executed often and shall not be computationally heavy.
See examples for \code{createLocalRepo} for some applications. 
If \code{connector=NULL} then information about connectors to an external database will be removed.
Note that it's an experimental feature.}

\item{user}{While working with the Remote repository. A character containing
a name of the Remote user that we want to make default.}

\item{repo}{While working with the Remote repository. A character containing
a name of the Remote repository that we want to make default.}

\item{branch}{While working with the Remote repository. A character containing a name of 
the Remote Repository's branch that we want to make default. Default \code{branch} is \code{master}.}

\item{subdir}{While working with the Remote repository. A character containing a name
of the Repository's directory on Remote that we want to make default.
If the Repository is stored in the main folder on the Remote repository,
this should be set to \code{subdir = "/"} as default.}

\item{repoType}{A character containing a type of the remote repository. 
Currently it can be 'github' or 'bitbucket'.}
}
\description{
\code{setLocalRepo} sets local \link{Repository}'s global path.
\code{setRemoteRepo} similarly sets Remote Repository's path.
See examples.
}
\details{
If you are working on a local Repository and you are tired of specifying \code{repoDir}
parameter in every function call that uses this parameter, you can set Repository's path
globally using \code{setLocalRepo} function and omit \code{repoDir} parameter
in future function calls.

If you are working on the Remote Repository and you are tired of specifying \code{user}, 
\code{repo}, \code{branch} and \code{subdir} parameters in every function call
that uses these parameters, you can set Remote Repository's path globally using
\code{setRemoteRepo} function and omit \code{user}, \code{repo}, \code{branch}
and \code{subdir} parameters in future function calls. See examples.

For local repositories, in this way, in the following function calls: 
\link{loadFromLocalRepo},\link{searchInLocalRepo}, \link{rmFromLocalRepo}, \link{zipLocalRepo}, 
\link{addTagsRepo}, \link{shinySearchInLocalRepo},
\link{getTagsLocal}, \link{showLocalRepo}, \link{summaryLocalRepo} 
\code{repoDir} parameter may be omitted. 
For remote repositories, in this way,
in the following function calls:
\link{zipRemoteRepo}, \link{loadFromRemoteRepo}, \link{searchInRemoteRepo},
\link{getTagsRemote}, \link{showRemoteRepo}, \link{summaryRemoteRepo},
\link{copyRemoteRepo}
parameters \code{user}, \code{repo}, \code{branch}, \code{subdir}  may be omitted.
}
\section{Contact}{

 
Bug reports and feature requests can be sent to 
\href{https://github.com/pbiecek/archivist/issues}{https://github.com/pbiecek/archivist/issues}
}

\examples{
\dontrun{
## Local version
exampleRepoDir <- tempfile()
createLocalRepo(repoDir = exampleRepoDir)
setLocalRepo(exampleRepoDir) 

data(iris)
data(swiss)

# From this moment repoDir parameter may be ommitted in the following functions
saveToRepo(iris)
saveToRepo(swiss) 
showLocalRepo()
showLocalRepo(method = "tags")
iris2 <- loadFromLocalRepo( "ff575c2" , value = TRUE)
searchInLocalRepo("name:i", fixed = FALSE)
getTagsLocal("ff575c261c949d073b2895b05d1097c3")
rmFromLocalRepo("4c43f")
showLocalRepo()
summaryLocalRepo()

# REMEMBER that in deleteLocalRepo you MUST specify repoDir parameter!
# deleteRepo doesn't take setLocalRepo's settings into consideration
deleteLocalRepo( exampleRepoDir, deleteRoot=TRUE)
rm( exampleRepoDir )

## Github version
setRemoteRepo( user="MarcinKosinski", repo="Museum", branch="master",
               subdir="ex1" )
               
# From this moment user, repo, branch, subdir parameters may be ommitted
# in the following functions:
showRemoteRepo()
loadFromRemoteRepo( "ff575c261c949d073b2895b05d1097c3")
this <- loadFromRemoteRepo( "ff", value = TRUE)
zipRemoteRepo()
file.remove(file.path(getwd(), "repository.zip")) # We can remove this zip file
searchInRemoteRepo( "name:", fixed= FALSE)
getTagsRemote("ff575c261c949d073b2895b05d1097c3")
summaryRemoteRepo( )

# To use multisearchInRemoteRepo we should use repository with more than one artifact. 
setRemoteRepo( user="pbiecek", repo="archivist"  )

# From this moment user and repo parameters may be ommitted in the following functions
showRemoteRepo()
searchInRemoteRepo( pattern=c("varname:Sepal.Width", "class:lm", "name:myplot123"), 
                         intersect = FALSE )
}
}
\references{
Biecek P and Kosinski M (2017). "archivist: An R Package for Managing,
Recording and Restoring Data Analysis Results." _Journal of Statistical
Software_, *82*(11), pp. 1-28. doi: 10.18637/jss.v082.i11 (URL:
http://doi.org/10.18637/jss.v082.i11).
URL https://github.com/pbiecek/archivist
}
\seealso{
\href{https://github.com/pbiecek/archivist/wiki}{https://github.com/pbiecek/archivist/wiki}

Other archivist: \code{\link{Repository}},
  \code{\link{Tags}}, \code{\link{\%a\%}},
  \code{\link{addHooksToPrint}}, \code{\link{addTagsRepo}},
  \code{\link{aformat}}, \code{\link{ahistory}},
  \code{\link{alink}}, \code{\link{aoptions}},
  \code{\link{archivist-package}},
  \code{\link{areadLocal}}, \code{\link{aread}},
  \code{\link{asearchLocal}}, \code{\link{asearch}},
  \code{\link{asession}}, \code{\link{atrace}},
  \code{\link{cache}}, \code{\link{copyLocalRepo}},
  \code{\link{createLocalRepo}},
  \code{\link{createMDGallery}},
  \code{\link{deleteLocalRepo}},
  \code{\link{getRemoteHook}}, \code{\link{getTagsLocal}},
  \code{\link{loadFromLocalRepo}}, \code{\link{md5hash}},
  \code{\link{removeTagsRepo}}, \code{\link{restoreLibs}},
  \code{\link{rmFromLocalRepo}},
  \code{\link{saveToLocalRepo}},
  \code{\link{searchInLocalRepo}},
  \code{\link{shinySearchInLocalRepo}},
  \code{\link{showLocalRepo}},
  \code{\link{splitTagsLocal}},
  \code{\link{summaryLocalRepo}},
  \code{\link{zipLocalRepo}}

Other archivist: \code{\link{Repository}},
  \code{\link{Tags}}, \code{\link{\%a\%}},
  \code{\link{addHooksToPrint}}, \code{\link{addTagsRepo}},
  \code{\link{aformat}}, \code{\link{ahistory}},
  \code{\link{alink}}, \code{\link{aoptions}},
  \code{\link{archivist-package}},
  \code{\link{areadLocal}}, \code{\link{aread}},
  \code{\link{asearchLocal}}, \code{\link{asearch}},
  \code{\link{asession}}, \code{\link{atrace}},
  \code{\link{cache}}, \code{\link{copyLocalRepo}},
  \code{\link{createLocalRepo}},
  \code{\link{createMDGallery}},
  \code{\link{deleteLocalRepo}},
  \code{\link{getRemoteHook}}, \code{\link{getTagsLocal}},
  \code{\link{loadFromLocalRepo}}, \code{\link{md5hash}},
  \code{\link{removeTagsRepo}}, \code{\link{restoreLibs}},
  \code{\link{rmFromLocalRepo}},
  \code{\link{saveToLocalRepo}},
  \code{\link{searchInLocalRepo}},
  \code{\link{shinySearchInLocalRepo}},
  \code{\link{showLocalRepo}},
  \code{\link{splitTagsLocal}},
  \code{\link{summaryLocalRepo}},
  \code{\link{zipLocalRepo}}

Other archivist: \code{\link{Repository}},
  \code{\link{Tags}}, \code{\link{\%a\%}},
  \code{\link{addHooksToPrint}}, \code{\link{addTagsRepo}},
  \code{\link{aformat}}, \code{\link{ahistory}},
  \code{\link{alink}}, \code{\link{aoptions}},
  \code{\link{archivist-package}},
  \code{\link{areadLocal}}, \code{\link{aread}},
  \code{\link{asearchLocal}}, \code{\link{asearch}},
  \code{\link{asession}}, \code{\link{atrace}},
  \code{\link{cache}}, \code{\link{copyLocalRepo}},
  \code{\link{createLocalRepo}},
  \code{\link{createMDGallery}},
  \code{\link{deleteLocalRepo}},
  \code{\link{getRemoteHook}}, \code{\link{getTagsLocal}},
  \code{\link{loadFromLocalRepo}}, \code{\link{md5hash}},
  \code{\link{removeTagsRepo}}, \code{\link{restoreLibs}},
  \code{\link{rmFromLocalRepo}},
  \code{\link{saveToLocalRepo}},
  \code{\link{searchInLocalRepo}},
  \code{\link{shinySearchInLocalRepo}},
  \code{\link{showLocalRepo}},
  \code{\link{splitTagsLocal}},
  \code{\link{summaryLocalRepo}},
  \code{\link{zipLocalRepo}}
}
\author{
Marcin Kosinski, \email{m.p.kosinski@gmail.com}

Przemyslaw Biecek, \email{przemyslaw.biecek@gmail.com}
}
\concept{archivist}
