% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setRepo.R
\name{setLocalRepo}
\alias{setLocalRepo}
\alias{setRemoteRepo}
\title{Set Repository's Global Path}
\usage{
setLocalRepo(repoDir)

setRemoteRepo(user, repo, branch = "master", subdir = "/",
  repoType = "github")
}
\arguments{
\item{repoDir}{A character denoting a directory of a Repository that we want to
make dafault.}

\item{user}{While working with the Remote repository. A character containing
a name of the Remote user that we want to make default.}

\item{repo}{While working with the Remote repository. A character containing
a name of the Remote repository that we want to make default.}

\item{branch}{While working with the Remote repository. A character containing a name of 
the Remote Repository's branch that we want to make default. Default \code{branch} is \code{master}.}

\item{subdir}{While working with the Remote repository. A character containing a name
of the Repository's directory on Remote that we want to make default.
If the Repository is stored in the main folder on the Remote repository,
this should be set to \code{subdir = "/"} as default.}

\item{repoType}{A character containing a type of the remote repository. 
Currently it can be 'github' or 'bitbucket'.}
}
\description{
\code{setLocalRepo} sets local \link{Repository}'s global path.
\code{setRemoteRepo} similarly sets Remote Repository's path.
See examples.
}
\details{
If you are working on a local Repository and you are tired of specifying \code{repoDir}
parameter in every function call that uses this parameter, you can set Repository's path
globally using \code{setLocalRepo} function and omit \code{repoDir} parameter
in future function calls.

If you are working on the Remote Repository and you are tired of specifying \code{user}, 
\code{repo}, \code{branch} and \code{subdir} parameters in every function call
that uses these parameters, you can set Remote Repository's path globally using
\code{setRemoteRepo} function and omit \code{user}, \code{repo}, \code{branch}
and \code{subdir} parameters in future function calls. See examples.

For local repositories, in this way, in the following function calls: 
\link{loadFromLocalRepo},\link{searchInLocalRepo}, \link{rmFromRepo}, \link{zipLocalRepo}, 
\link{multiSearchInLocalRepo}, \link{addTagsRepo}, \link{shinySearchInLocalRepo},
\link{getTagsLocal}, \link{showLocalRepo}, \link{summaryLocalRepo} 
\code{repoDir} parameter may be omitted. 
For remote repositories, in this way,
in the following function calls:
\link{zipRemoteRepo}, \link{loadFromRemoteRepo}, \link{searchInRemoteRepo},
\link{getTagsRemote}, \link{showRemoteRepo}, \link{summaryRemoteRepo},
\link{multiSearchInRemoteRepo}, \link{copyRemoteRepo}
parameters \code{user}, \code{repo}, \code{branch}, \code{subdir}  may be omitted.
}
\examples{

## Local version
exampleRepoDir <- tempfile()
createLocalRepo(repoDir = exampleRepoDir)
setLocalRepo(exampleRepoDir) 

data(iris)
data(swiss)

# From this moment repoDir parameter may be ommitted in the following functions
saveToRepo(iris)
saveToRepo(swiss) 
showLocalRepo()
showLocalRepo(method = "tags")
iris2 <- loadFromLocalRepo( "ff575c2" , value = TRUE)
searchInLocalRepo("name:i", fixed = F)
getTagsLocal("ff575c261c949d073b2895b05d1097c3")
rmFromRepo("4c43f")
showLocalRepo()
summaryLocalRepo()

# REMEMBER that in deleteLocalRepo you MUST specify repoDir parameter!
# deleteRepo doesn't take setLocalRepo's settings into consideration
deleteLocalRepo( exampleRepoDir, deleteRoot=TRUE)
rm( exampleRepoDir )

## Github version
setRemoteRepo( user="MarcinKosinski", repo="Museum", branch="master",
               subdir="ex1" )
               
# From this moment user, repo, branch, subdir parameters may be ommitted
# in the following functions:
showRemoteRepo()
loadFromRemoteRepo( "ff575c261c949d073b2895b05d1097c3")
this <- loadFromRemoteRepo( "ff", value = T)
zipRemoteRepo()
file.remove(file.path(getwd(), "repository.zip")) # We can remove this zip file
searchInRemoteRepo( "name:", fixed= FALSE)
getTagsRemote("ff575c261c949d073b2895b05d1097c3")
summaryRemoteRepo( )

# To use multisearchInRemoteRepo we should use repository with more than one artifact. 
setRemoteRepo( user="pbiecek", repo="archivist"  )

# From this moment user and repo parameters may be ommitted in the following functions
showRemoteRepo()
multiSearchInRemoteRepo( patterns=c("varname:Sepal.Width", "class:lm", "name:myplot123"), 
                         intersect = FALSE )
}
\author{
Marcin Kosinski , \email{m.p.kosinski@gmail.com}

Przemyslaw Biecek, \email{przemyslaw.biecek@gmail.com}
}
\seealso{
\href{https://github.com/pbiecek/archivist/wiki}{https://github.com/pbiecek/archivist/wiki}

Other archivist: \code{\link{Repository}},
  \code{\link{Tags}}, \code{\link{\%a\%}},
  \code{\link{addHooksToPrint}}, \code{\link{addTagsRepo}},
  \code{\link{aformat}}, \code{\link{ahistory}},
  \code{\link{alink}}, \code{\link{aoptions}},
  \code{\link{archivist-package}}, \code{\link{aread}},
  \code{\link{asearch}}, \code{\link{asession}},
  \code{\link{cache}}, \code{\link{copyLocalRepo}},
  \code{\link{createLocalRepo}},
  \code{\link{createMDGallery}},
  \code{\link{deleteLocalRepo}},
  \code{\link{getRemoteHook}}, \code{\link{getTagsLocal}},
  \code{\link{loadFromLocalRepo}}, \code{\link{md5hash}},
  \code{\link{rmFromRepo}}, \code{\link{saveToRepo}},
  \code{\link{searchInLocalRepo}},
  \code{\link{shinySearchInLocalRepo}},
  \code{\link{showLocalRepo}},
  \code{\link{splitTagsLocal}},
  \code{\link{summaryLocalRepo}},
  \code{\link{zipLocalRepo}}

Other archivist: \code{\link{Repository}},
  \code{\link{Tags}}, \code{\link{\%a\%}},
  \code{\link{addHooksToPrint}}, \code{\link{addTagsRepo}},
  \code{\link{aformat}}, \code{\link{ahistory}},
  \code{\link{alink}}, \code{\link{aoptions}},
  \code{\link{archivist-package}}, \code{\link{aread}},
  \code{\link{asearch}}, \code{\link{asession}},
  \code{\link{cache}}, \code{\link{copyLocalRepo}},
  \code{\link{createLocalRepo}},
  \code{\link{createMDGallery}},
  \code{\link{deleteLocalRepo}},
  \code{\link{getRemoteHook}}, \code{\link{getTagsLocal}},
  \code{\link{loadFromLocalRepo}}, \code{\link{md5hash}},
  \code{\link{rmFromRepo}}, \code{\link{saveToRepo}},
  \code{\link{searchInLocalRepo}},
  \code{\link{shinySearchInLocalRepo}},
  \code{\link{showLocalRepo}},
  \code{\link{splitTagsLocal}},
  \code{\link{summaryLocalRepo}},
  \code{\link{zipLocalRepo}}
}

