% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitTags.R
\name{splitTagsLocal}
\alias{splitTags}
\alias{splitTagsGithub}
\alias{splitTagsLocal}
\alias{splitTagsRemote}
\title{Split Tags in Repository}
\usage{
splitTagsLocal(repoDir = aoptions("repoDir"))

splitTagsRemote(repo = aoptions("repo"), user = aoptions("user"),
  branch = aoptions("branch"), subdir = aoptions("subdir"),
  repoType = aoptions("repoType"))

splitTagsGithub(repo = aoptions("repo"), user = aoptions("user"),
  branch = aoptions("branch"), subdir = aoptions("subdir"),
  repoType = aoptions("repoType"))

splitTags(repoDir = NULL, repo = aoptions("repo"),
  user = aoptions("user"), branch = "master", subdir = aoptions("subdir"),
  repoType = aoptions("repoType"), local = TRUE)
}
\arguments{
\item{repoDir}{While working with the local repository. A character denoting 
an existing directory of the Repository. If it is set to \code{NULL} (by default),
it will use the \code{repoDir} specified in \link{setLocalRepo}.}

\item{repo}{While working with the Github repository. A character containing
a name of the Github repository on which the Repository is stored.
By default set to \code{NULL} - see \code{Note}.}

\item{user}{While working with the Github repository. A character containing
a name of the Github user on whose account the \code{repo} is created.
By default set to \code{NULL} - see \code{Note}.}

\item{branch}{While working with the Github repository. A character containing
a name of the Github Repository's branch on which the Repository is stored.
Default \code{branch} is \code{master}.}

\item{subdir}{While working with the Github repository. A character containing
a name of a directory on the Github repository on which the Repository is stored.
If the Repository is stored in the main folder of the Github repository,
this should be set to \code{subdir = "/"} as default.}

\item{repoType}{A character containing a type of the remote repository. 
Currently it can be 'github' or 'bitbucket'.}
}
\value{
A \code{data.frame} with 4 columns: \code{artifact}, \code{tagKey},
\code{tagValue} and \code{createdDate}, corresponding to the current state of \link{Repository}.
}
\description{
\code{splitTagsLocal} and \code{splitTagsGithub} functions split \code{tag} column from
\emph{tag} table placed in \code{backpack.db} into two separate columns:
\code{tagKey} and \code{tagValue}.
}
\details{
\code{tag} column from \emph{tag} table has normally the follwing structure:
\code{TagKey:TagValue}. \code{splitTagsLocal} and \code{splitTagsGithub} functions
can be used to split \code{tag} column into two separate columns:
\code{tagKey} and \code{tagValue}. As a result functions from \code{dplyr} package
can be used to easily summarize, search, and extract artifacts' Tags.
See \code{examples}.
}
\note{
If \code{repo} and \code{user} are set to \code{NULL} (as default) in the Github mode
then global parameters set in \link{setGithubRepo} function are used.

Sometimes we can use \code{addTags*} function or \code{userTags} parameter
in \code{saveToRepo} to specify a \code{Tag} which might not match
\code{TagKey:TagValue} structure. It is simply \code{Tag}. In this case 
\code{tagKey = userTags} and \code{tagValue = Tag}. See \code{examples}.

To learn more about \code{Tags} and \code{Repository} structure check 
\link{Tags} and \link{Repository}.
}
\examples{
\dontrun{
## LOCAL VERSION 

setLocalRepo(system.file("graphGallery", package = "archivist"))
head(showLocalRepo(method = "tags"))
head(splitTagsLocal() )

## Github Version
# Let's check how does table tag look like while we are using the
# Gitub repository.
# We will choose only special columns of data frames that show Tags
head(showGithubRepo( user = "pbiecek", repo = "archivist", method = "tags" )[,2])
head(splitTagsGithub( user = "pbiecek", repo = "archivist" )[,2:3])

head(splitTagsRemote("PieczaraPietraszki", "BetaAndBit", "master", "UniwersytetDzieci/arepo"))
}
}
\author{
Witold Chodor , \email{witoldchodor@gmail.com}
}
\seealso{
Other archivist: \code{\link{Repository}},
  \code{\link{Tags}}, \code{\link{\%a\%}},
  \code{\link{addHooksToPrint}}, \code{\link{addTagsRepo}},
  \code{\link{aformat}}, \code{\link{ahistory}},
  \code{\link{alink}}, \code{\link{aoptions}},
  \code{\link{archivist-package}}, \code{\link{aread}},
  \code{\link{asearch}}, \code{\link{asession}},
  \code{\link{cache}}, \code{\link{copyLocalRepo}},
  \code{\link{createEmptyLocalRepo}},
  \code{\link{createMDGallery}},
  \code{\link{deleteLocalRepo}},
  \code{\link{getRemoteHook}}, \code{\link{getTagsLocal}},
  \code{\link{loadFromLocalRepo}}, \code{\link{md5hash}},
  \code{\link{rmFromRepo}}, \code{\link{saveToRepo}},
  \code{\link{searchInLocalRepo}},
  \code{\link{setLocalRepo}},
  \code{\link{shinySearchInLocalRepo}},
  \code{\link{showLocalRepo}},
  \code{\link{summaryLocalRepo}},
  \code{\link{zipLocalRepo}}
}

