% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{Tags}
\alias{Tags}
\title{Tags}
\description{
\code{Tags} are attributes of an object, i.e., a class, a name, names of object's parts, etc.. The list of object tags vary across object's classes.
}
\details{
\code{Tags} are attributes of an object. \code{Tags} can be an object's \code{name}, \code{class} or \code{archiving date}.
Furthermore, for various object's classes more different \code{Tags} are available and can
be searched by \link{searchInLocalRepo} or \link{searchInGithubRepo} functions.

\code{Tags} are stored in the \link{Repository}. If data is extracted from object a special \code{Tag}
named \code{relationWith} is created, and specifies with which object this data is related to.

So far supported objects list is presented below. Objects are divided thematically.
The newest list is also available on \pkg{archivist} \code{wikis} on
\href{https://github.com/pbiecek/archivist/wiki/archivist-package---Tags}{https://github.com/pbiecek/archivist/wiki/archivist-package---Tags}.

Regression Models

\describe{
  \item{\code{lm}}{
  \itemize{
     \item coefname
     \item class
     \item call
     \item name
     \item data
     \item date
  }
  }
  \item{\code{glmnet}}{
  \itemize{
     \item date
     \item name
     \item class
     \item call
     \item beta
     \item lambda
  }
  }
  \item{\code{survfit}}{
  \itemize{
     \item date
     \item name
     \item class
     \item call
     \item strata
     \item type
  }
  }
}
Plots

\describe{
  \item{\code{ggplot}}{
  \itemize{
     \item name
     \item class
     \item date
     \item data
     \item labelx
     \item labely
  }
  }
  \item{\code{trellis}}{
  \itemize{
     \item date
     \item name
     \item class
     \item call
  }
  }
}
Results of Agglomeration Methods

\describe{
\item{\code{twins which is a result of agnes, diana or mona functions}}{
  \itemize{
     \item date
     \item name
     \item class
     \item ac
     \item merge
     \item data
  }
  }
  \item{\code{partition which is a result of pam, clara or fanny functions}}{
  \itemize{
     \item date
     \item name
     \item class
     \item call
     \item data
     \item objective
  }
  }
  \item{\code{lda}}{
  \itemize{
     \item date
     \item name
     \item class
     \item call
  }
  }
  \item{\code{qda}}{
  \itemize{
     \item date
     \item name
     \item class
     \item call
     \item terms
  }
  }
}
Statistical Tests

\describe{
  \item{\code{htest}}{
  \itemize{
     \item alternative
     \item method
     \item date
     \item name
     \item class
  }
  }
}

When non of above is specified, tags are corresponded by default

\describe{
  \item{\code{default}}{
  \itemize{
     \item name
     \item class
     \item date
  }
  }
  \item{\code{data.frame}}{
  \itemize{
     \item name
     \item class
     \item date
     \item varname
  }
  }
  }
}
\seealso{
Functions using \code{Tags} are \link{searchInLocalRepo} and \link{searchInGithubRepo}.

Other archivist: \code{\link{Repository}};
  \code{\link{archivist-package}};
  \code{\link{createEmptyRepo}}; \code{\link{deleteRepo}};
  \code{\link{infoGithubRepo}},
  \code{\link{infoLocalRepo}};
  \code{\link{loadFromGithubRepo}},
  \code{\link{loadFromLocalRepo}}; \code{\link{md5hash}};
  \code{\link{rmFromRepo}}; \code{\link{saveToRepo}};
  \code{\link{searchInGithubRepo}},
  \code{\link{searchInLocalRepo}};
  \code{\link{summaryGithubRepo}},
  \code{\link{summaryLocalRepo}}
}

